% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gotm.r
\name{AIC.gotm}
\alias{AIC.gotm}
\title{Extracts Akaike Information Criterion from the fitted model}
\usage{
AIC.gotm(object, ..., k = 2L)
}
\arguments{
\item{object}{\code{gotm} object.}

\item{...}{additional objects of the same type.}

\item{k}{numeric, the penalty per parameter to be used; the default k = 2 is the classical AIC.}
}
\description{
Extracts Akaike Information Criterion from the fitted model
}
\author{
Maciej J. Danko <\email{danko@demogr.mpg.de}> <\email{maciej.danko@gmail.com}>
}
\keyword{internal}

