% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{AIC.hopit}
\alias{AIC.hopit}
\title{Extracts Akaike Information Criterion from the fitted model}
\usage{
AIC.hopit(object, ..., k = 2L)
}
\arguments{
\item{object}{\code{hopit} object.}

\item{...}{additional objects of the same type.}

\item{k}{numeric, the penalty per parameter to be used; the default k = 2 is the classical AIC.}
}
\description{
Extracts Akaike Information Criterion from the fitted model
}
\author{
Maciej J. Danko
}
\keyword{internal}
