% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gotm.r
\name{anova.gotm}
\alias{anova.gotm}
\title{Likelihood Ratio Tests}
\usage{
\method{anova}{gotm}(object, ..., method = c("sequential", "with.first"),
  direction = c("decreasing", "increasing"))
}
\description{
Likelihood Ratio Tests
}
\author{
Maciej J. Danko <\email{danko@demogr.mpg.de}> <\email{maciej.danko@gmail.com}>
}
\keyword{internal}

