% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{anova.hopit}
\alias{anova.hopit}
\title{LRT Tables}
\usage{
\method{anova}{hopit}(object, ..., method = c("sequential", "with.first"),
direction = c("decreasing", "increasing"))
}
\arguments{
\item{object}{an object containing the results returned by a \code{hopit}.}

\item{...}{additional objects of the same type.}

\item{method}{the method of model comparison. Choose \code{"sequential"} for 1-2, 2-3, 3-4, ... comparisons or
\code{"with.first"} for 1-2, 1-3, 1-4, ... comparisons.}

\item{direction}{determine if complexity of listed models is \code{"increasing"} or \code{"decreasing"} (default).}
}
\value{
\code{lrt.hopit} or \code{anova.hopit} objects.
}
\description{
Compute likelihood rato test for two or more \code{hopit} objecs.
}
\seealso{
\code{\link{print.anova.hopit}}, \code{\link{print.lrt.hopit}}, \code{\link{lrt.hopit}}, \code{\link{hopit}}.
}
\author{
Maciej J. Danko
}
