% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportingstyles.r
\name{basiccutpoints}
\alias{basiccutpoints}
\title{Calcualte threshold cut-points using Jurges' method}
\usage{
basiccutpoints(model, subset = NULL, plotf = TRUE, mar = c(4, 4, 1, 1),
  oma = c(0, 0, 0, 0), revf = NULL)
}
\arguments{
\item{model}{fitted gotm model.}

\item{subset}{an optional vector specifying a subset of observations.}

\item{plotf}{logical indicating if to plot the results.}

\item{mar}{see \code{\link{par}}}

\item{oma}{see \code{\link{par}}}

\item{revf}{logical indicating if self-reported health classes are ordered in increasing order.}
}
\description{
Calcualte threshold cut-points using Jurges' method.
}
\keyword{internal}

