% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot.R
\name{boot_hopit_CI}
\alias{boot_hopit_CI}
\title{Calculating confidence intervals of a boot object using percentile method}
\usage{
boot_hopit_CI(boot, alpha = 0.05, bounds = c("both", "lo", "up"))
}
\arguments{
\item{boot}{a boot object calculated by \code{\link{boot_hopit}} .}

\item{alpha}{significance level.}

\item{bounds}{one of \code{"both"}, \code{"lo"}, \code{"up"}.}
}
\description{
Calculating confidence intervals of a boot object using percentile method
}
\examples{
# see examples in boot_hopit() function.
}
\seealso{
\code{\link{boot_hopit}}, \code{\link{getLevels}}, \code{\link{getCutPoints}}, \code{\link{latentIndex}}, \code{\link{standardiseCoef}}, \code{\link{hopit}}.
}
\author{
Maciej J. Danko
}
