% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot.R
\name{boot_latent_hopit}
\alias{boot_latent_hopit}
\title{Bootstraping hopit model}
\usage{
boot_latent_hopit(model, data, func, nboot = 500, ...)
}
\arguments{
\item{model}{a fitted \code{Hopit} model.}

\item{data}{data used to fit the model}

\item{func}{function to be bootstrapped of the form func(model, data, ...).}

\item{nboot}{number of bootstrap replicates.}
}
\description{
Bootstraping hopit model
}
\author{
Maciej J. Danko
}
