% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{calcYYY}
\alias{calcYYY}
\title{INTERNAL: Calculate special matrices for gradient calaculation}
\usage{
calcYYY(model)
}
\arguments{
\item{model}{fitted model.}
}
\value{
updated model
}
\description{
INTERNAL: Calculate special matrices for gradient calaculation
}
\author{
Maciej J. Danko
}
\keyword{internal}
