% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surveyclone.R
\name{clone.of.htvar.matrix}
\alias{clone.of.htvar.matrix}
\title{survey:::htvar.matrix clone}
\usage{
clone.of.htvar.matrix(xcheck, Dcheck)
}
\arguments{
\item{xcheck, Dcheck}{internal parameters.}
}
\description{
survey:::htvar.matrix clone
}
\details{
clone taken from the survey v3.35 package.
}
\author{
Thomas Lumley
}
\keyword{internal}
