% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surveyclone.R
\name{clone.of.ppsvar}
\alias{clone.of.ppsvar}
\title{survey:::ppsvar clone}
\usage{
clone.of.ppsvar(x, design)
}
\arguments{
\item{x, design}{internal parameters.}
}
\description{
survey:::ppsvar clone
}
\details{
clone taken from the survey v3.35 package.
}
\author{
Thomas Lumley
}
\keyword{internal}
