% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surveyclone.R
\name{clone.of.ygvar.matrix}
\alias{clone.of.ygvar.matrix}
\title{survey:::ygvar.matrix clone}
\usage{
clone.of.ygvar.matrix(xcheck, Dcheck)
}
\arguments{
\item{xcheck, Dcheck}{internal parameters.}
}
\description{
survey:::ygvar.matrix clone
}
\details{
clone taken from the survey v3.35 package.
}
\author{
Thomas Lumley
}
\keyword{internal}
