% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gotm_imprv.r
\name{coef.gotm}
\alias{coef.gotm}
\alias{coefficients.gotm}
\title{Extracting coefficients of fitted \code{gotm} object}
\usage{
\method{coef}{gotm}(object, aslist = FALSE, ...)
}
\arguments{
\item{object}{\code{gotm} object.}

\item{aslist}{logical indicating if model coefficients should be returned as a list of three vectors
related to latent variable, threshold lambdas, and threshold gammas.}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Extracting coefficients of fitted \code{gotm} object
}
\author{
Maciej J. Danko
}
\keyword{internal}
