% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{coef.hopit}
\alias{coef.hopit}
\alias{coefficients.hopit}
\title{Extracting model coefficients}
\usage{
\method{coef}{hopit}(object, aslist = FALSE, ...)
}
\arguments{
\item{object}{an \code{hopit} object.}

\item{aslist}{logical indicating if model coefficients should be returned
as a list of three vectors
related to latent variable, threshold lambdas, and threshold gammas.}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Extract model coefficients from \code{hopit} model.
}
\author{
Maciej J. Danko
}
\keyword{internal}
