% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gotm_imprv.r
\name{col_path}
\alias{col_path}
\title{INTERNAL: Column path in a matrix}
\usage{
col_path(mat, y, offset = 0)
}
\arguments{
\item{mat}{a matrix}

\item{y}{a vector with column indices corresponding to each row in the matrix \code{mat}.}
}
\description{
For each row of a matrix \code{mat} extracts a value corresponding to a column stored in a vector \code{y}.
}
\author{
Maciej J. Danko
}
\keyword{internal}
