% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportingstyles.r
\name{comparehealthlevels}
\alias{comparehealthlevels}
\title{Plot a comparison of original and adjusted health levels}
\usage{
comparehealthlevels(object, pch = 19, xlab = "Original frequency [\%]",
  ylab = "Adjusted frequency [\%]", mar = c(2.5, 1, 1.5, 1), oma = c(4, 4,
  0.1, 0.1), ratio = 1)
}
\arguments{
\item{object}{an object generated by \code{\link{gethealthlevels}}}

\item{pch, xlab, ylab, mar, oma}{common graphical parameters.}

\item{ratio}{an aspect ratio for panels composition.}
}
\description{
Plot a comparison of original and adjusted health levels
}

