% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gotm.r
\name{cumsum_row}
\alias{cumsum_row}
\title{INTERNAL: Do cumsum() in each row of the matrix}
\usage{
cumsum_row(mat)
}
\arguments{
\item{mat}{a matrix.}
}
\description{
INTERNAL: Do cumsum() in each row of the matrix
}
\author{
Maciej J. Danko <\email{danko@demogr.mpg.de}> <\email{maciej.danko@gmail.com}>
}
\keyword{internal}

