% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{data2freq}
\alias{data2freq}
\title{Convert individual data to frequency table of unique combination of dependent and independent variables}
\usage{
data2freq(formula, data, FreqNam = "Freq")
}
\arguments{
\item{formula}{formula indicating, which variables will be used to construct new database.}

\item{data}{data.frame including all variables listed in formula.}

\item{FreqNam}{name of the column with frequencies.}
}
\description{
Convert individual data to frequency table of unique combination of dependent and independent variables
}
\author{
Maciej J. Danko
}
