% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportingstyles.r
\name{disabilityweights}
\alias{disabilityweights}
\title{Calculate disability weights}
\usage{
disabilityweights(model, plotf = TRUE, mar = c(15, 4, 1, 1), oma = c(0, 0,
  0, 0))
}
\description{
Calculate disability weights.
}

