% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gotm_imprv.r
\name{DummyMat2Vector}
\alias{DummyMat2Vector}
\title{INTERNAL: Converts a matrix with dummies in columns into categorical vector}
\usage{
DummyMat2Vector(D)
}
\arguments{
\item{D}{a matrix of dummies.}
}
\description{
INTERNAL: Converts a matrix with dummies in columns into categorical vector
}
\author{
Maciej J. Danko  <\email{maciej.danko@gmail.com}>
}
\keyword{internal}
