% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gotm_imprv.r
\name{fit.vglm}
\alias{fit.vglm}
\title{INTERNAL: Fit vglm to get parameters of the model}
\usage{
fit.vglm(model, data, start = NULL)
}
\description{
INTERNAL: Fit vglm to get parameters of the model
}
\keyword{internal}
