% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get.hopit.start}
\alias{get.hopit.start}
\title{INTERNAL: Get starting parameters}
\usage{
get.hopit.start(object, data)
}
\arguments{
\item{object}{an \code{hopit} object.}

\item{data}{data.frame with data used to fit the object.}
}
\value{
an updated \code{hopit} object.
}
\description{
INTERNAL: Get starting parameters
}
\author{
Maciej J. Danko
}
\keyword{internal}
