% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gotm.r
\name{get.start.gotm}
\alias{get.start.gotm}
\title{Get starting parameters from less or more complicated hierarchical models}
\usage{
get.start.gotm(object, reg.formula, thresh.formula, data, asList = FALSE)
}
\description{
Get starting parameters from less or more complicated hierarchical models
}

