% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hopit.R
\name{get.start.hopit}
\alias{get.start.hopit}
\title{Get starting parameters from less or more complicated hierarchical models}
\usage{
get.start.hopit(object, reg.formula, thresh.formula, data, asList = FALSE)
}
\description{
Get starting parameters from less or more complicated hierarchical models
}
