% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gotm_imprv.r
\name{get.vglm.start}
\alias{get.vglm.start}
\title{INTERNAL: Fit vglm to get parameters of the model}
\usage{
get.vglm.start(model, data, thresh.start = model$thresh.start)
}
\arguments{
\item{model}{\code{gotm} object.}

\item{data}{data.frame with data used to fit the model.}
}
\description{
INTERNAL: Fit vglm to get parameters of the model
}
\author{
Maciej J. Danko
}
\keyword{internal}
