% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportingstyles.R
\name{gethealthlevels}
\alias{gethealthlevels}
\title{Calcualte adjusted health levels.}
\usage{
gethealthlevels(model, formula = model$thresh.formula,
  data = environment(model$thresh.formula), revf = NULL,
  sort.flag = FALSE, plotf = TRUE, sep = "_", mar = c(7, 2, 1.5, 0.5),
  oma = c(0, 3, 0, 0))
}
\arguments{
\item{model}{a fitted \code{hopit} model.}

\item{formula}{a formula containing the variables. It is by default set to threshold formula.}

\item{data}{data used to fit the model.}

\item{revf}{logical indicating if self-reported health classes are ordered in increasing order.}

\item{plotf}{logical indicating if to plot the results.}

\item{sep}{separator for levels names.}

\item{mar}{see \code{\link{par}}.}

\item{oma}{see \code{\link{par}}.}
}
\description{
Calcualte adjusted health levels according to th Jurges' method.
}
