% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot.R
\name{gethealthlevels_boot}
\alias{gethealthlevels_boot}
\title{Bootstraping health levels}
\usage{
gethealthlevels_boot(model, formula = model$thresh.formula,
  data = environment(model$thresh.formula), revf = NULL, nboot = 1000,
  alpha = 0.05, plotF = FALSE, namefunc = identity, col = c("red4",
  "blue4"), pch = c(1, 0))
}
\description{
Bootstraping health levels
}
\author{
Maciej J. Danko
}
