% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gotm.r
\name{gotm.design}
\alias{gotm.design}
\title{Auxiliary for setting a simple survey design for \code{gotm}}
\usage{
gotm.design(PWeights = NULL, FWeights = NULL, PSU = NULL)
}
\arguments{
\item{PWeights}{Probability weights (the inverse of \code{FWeights})}

\item{FWeights}{Probability of an observation being selected into the sample.
Either \code{PWeight} or \code{FWeight} can be delivered (but not both simultaneously)}

\item{PSU}{Identificator of the PSU unit. Each P- and F- weight should correspond to exactly one PSU.}
}
\description{
Auxiliary for setting a simple survey design for \code{gotm}
}
\author{
Maciej J. Danko <\email{danko@demogr.mpg.de}> <\email{maciej.danko@gmail.com}>
}

