% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{gotm_c_link}
\alias{gotm_c_link}
\title{INTERNAL: Decode link parmeter}
\usage{
gotm_c_link(model)
}
\arguments{
\item{model}{fitted model.}
}
\value{
updated model
}
\description{
INTERNAL: Decode link parmeter
}
\keyword{internal}
