% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gotm_imprv.r
\name{gotm_derivLL}
\alias{gotm_derivLL}
\title{INTERNAL: The gradient of the log likelihood function}
\usage{
gotm_derivLL(parameters = model$coef, model,
  thresh.start = model$thresh.start, collapse = TRUE, use_weights = TRUE,
  negative = FALSE)
}
\description{
INTERNAL: The gradient of the log likelihood function
}
\author{
Maciej J. Danko
}
\keyword{internal}
