% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gotm_imprv.r
\name{gotm_fitter}
\alias{gotm_fitter}
\title{INTERNAL: Fit \code{gotm}}
\usage{
gotm_fitter(model, start = model$start, thresh.start = model$thresh.start,
  use_weights = TRUE)
}
\arguments{
\item{model}{\code{gotm} object}

\item{start}{starting parameters}

\item{control}{a list with control parameters. See \code{\link{gotm.control}}.}
}
\description{
Fit the model.
}
\author{
Maciej J. Danko
}
\keyword{internal}
