% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gotm.r
\name{gotm_GetInitial}
\alias{gotm_GetInitial}
\title{INTERNAL: Anticipate initial values of the model}
\usage{
gotm_GetInitial(model, data = NULL)
}
\arguments{
\item{model}{\code{gotm} object.}

\item{data}{data.frame with data used to fit the model.}
}
\description{
INTERNAL: Anticipate initial values of the model
}
\author{
Maciej J. Danko <\email{danko@demogr.mpg.de}> <\email{maciej.danko@gmail.com}>
}
\keyword{internal}

