% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gotm.r
\name{gotm_latentrange}
\alias{gotm_latentrange}
\title{INTERNAL: Calculate maximum possible latent range}
\usage{
gotm_latentrange(model, data)
}
\arguments{
\item{model}{a fitted \code{gotm} model.}

\item{data}{a data used to fit the model}
}
\description{
INTERNAL: Calculate maximum possible latent range
}
\keyword{internal}

