% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gotm_imprv.r
\name{gotm_negLL}
\alias{gotm_negLL}
\title{INTERNAL: The log likelihood function}
\usage{
gotm_negLL(parameters = model$coef, model,
  thresh.start = model$thresh.start, collapse = TRUE, use_weights = TRUE,
  negative = TRUE)
}
\description{
INTERNAL: The log likelihood function
}
\author{
Maciej J. Danko
}
\keyword{internal}
