% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gotm.r
\name{gotm_reg_thresh}
\alias{gotm_reg_thresh}
\title{INTERNAL: Calculation of cut-points (threshold)}
\usage{
gotm_reg_thresh(thresh.lambda, thresh.gamma, model)
}
\description{
INTERNAL: Calculation of cut-points (threshold)
}
\author{
Maciej J. Danko <\email{danko@demogr.mpg.de}> <\email{maciej.danko@gmail.com}>
}
\keyword{internal}

