% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gotm_imprv.r
\name{gotm_reg_thresh_old}
\alias{gotm_reg_thresh_old}
\title{INTERNAL: Calculation of cut-points (threshold), OLD versiom}
\usage{
gotm_reg_thresh_old(thresh.lambda, thresh.gamma, model)
}
\description{
INTERNAL: Calculation of cut-points (threshold), OLD versiom
}
\author{
Maciej J. Danko
}
\keyword{internal}
