% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gotm.r
\name{gotm_Threshold}
\alias{gotm_Threshold}
\title{INTERNAL: Calculation of cut-points (threshold)}
\usage{
gotm_Threshold(thresh.lambda, thresh.gamma, model = NULL, control = list(),
  N = model$N, J = model$J, offset = model$thresh.offset,
  mod.mat = model$thresh.mm, method = model$thresh.method,
  use.cov = !model$thresh.no.cov, extended.output = FALSE)
}
\description{
INTERNAL: Calculation of cut-points (threshold)
}
\author{
Maciej J. Danko <\email{danko@demogr.mpg.de}> <\email{maciej.danko@gmail.com}>
}
\keyword{internal}

