% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gotm_imprv.r
\name{gotm_Threshold_old}
\alias{gotm_Threshold_old}
\title{INTERNAL: Calculation of cut-points (threshold), OLD versiom}
\usage{
gotm_Threshold_old(thresh.lambda, thresh.gamma, model = NULL,
  extended.output = FALSE)
}
\description{
INTERNAL: Calculation of cut-points (threshold), OLD versiom
}
\author{
Maciej J. Danko
}
\keyword{internal}
