% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{\%c\%}
\alias{\%c\%}
\title{Check if one set is a subset of an another subset}
\usage{
x \%c\% y
}
\arguments{
\item{x, }{y numeric vectors}
}
\description{
Check if one set is a subset of an another subset
}
