% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{\%notc\%}
\alias{\%notc\%}
\title{Not \%c\% function}
\usage{
x \%notc\% y
}
\arguments{
\item{x, y}{numeric vectors}
}
\description{
Check whether the y set contains none of elements of the x set
}
\author{
Maciej J. Danko
}
