% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gotm.r
\name{HealthIndex.gotm}
\alias{HealthIndex.gotm}
\title{Calculate health index from the fitted \code{gotm} object}
\usage{
HealthIndex.gotm(model)
}
\arguments{
\item{model}{fitted \code{gotm} object.}
}
\description{
Calculate health index from the fitted \code{gotm} object
}

