% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/healthsurvey-data.R
\docType{data}
\name{healthsurvey}
\alias{healthsurvey}
\title{Artificially generated health survey data}
\format{A data frame with 10000 rows and 11 variables:
\describe{
  \item{ID}{Personal number}
  \item{health}{reported health, 5 levels}
  \item{hypertenssion}{has hypertenssion? Yes or No}
  \item{cholesterol}{has high cholesterol? Yes or No}
  \item{heart_stroke_respiratory}{has respiratory problems or had stroke or heart attack? Yes or No}
  \item{poor_mobility_or_grip}{has poor mobility or cannot perform grip strength? Yes or No}
  \item{depression}{has depression? Yes or No}
  \item{sex}{sex: woman or man}
  \item{ageclass}{categorized age: [50,60), [60,70), [70,80), [80,120)}
  \item{csw}{cross-sectional survey weights}
  \item{psu}{primary statistical unit}
}}
\source{
Data was randomly generated using probabilities of occurence of particular
combiantion of diseases, coditions, sex, age, and reported health.
These probabilities were obtained from the WAVE1 SHARE database
(DOIs: 10.6103/SHARE.w1.600), see Börsch-Supan et al for methodological details (Börsch-Supan et al. 2013).

None of the records represent a true individual.
}
\usage{
healthsurvey
}
\description{
A dataset containing artificial survey data
}
\references{
Börsch-Supan A, Brandt M, Hunkler C, et al (2013) Data resource profile: The survey of health, ageing and retirement in europe (share). Int J Epidemiol 42:992–1001. doi: 10.1093/ije/dyt088
}
\keyword{datasets}
