% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{hopit_c_link}
\alias{hopit_c_link}
\title{INTERNAL: Decode link parameter}
\usage{
hopit_c_link(model)
}
\arguments{
\item{model}{fitted model.}
}
\value{
updated model.
}
\description{
INTERNAL: Decode link parameter
}
\author{
Maciej J. Danko
}
\keyword{internal}
