% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hopit.R
\name{hopit_derivLL}
\alias{hopit_derivLL}
\title{INTERNAL: The gradient of the log likelihood function}
\usage{
hopit_derivLL(parameters = model$coef, model, collapse = TRUE,
  use_weights = TRUE, negative = FALSE)
}
\description{
INTERNAL: The gradient of the log likelihood function
}
\author{
Maciej J. Danko
}
\keyword{internal}
