% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hopit.R
\name{hopit_ExtractParameters}
\alias{hopit_ExtractParameters}
\title{INTERNAL: Extract model parameters in a form of list}
\usage{
hopit_ExtractParameters(model, parameters, parcount = model$parcount)
}
\arguments{
\item{model}{\code{hopit} object}

\item{parameters}{model parameters (optional). If not delivered then taken from \code{model$coef}}
}
\description{
Extract model parameters in a form of a list
}
\author{
Maciej J. Danko
}
\keyword{internal}
