% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hopit.R
\name{hopit_fitter}
\alias{hopit_fitter}
\title{INTERNAL: Fit \code{hopit}}
\usage{
hopit_fitter(model, start = model$start, use_weights = TRUE)
}
\arguments{
\item{model}{\code{hopit} object}

\item{start}{starting parameters}

\item{control}{a list with control parameters. See \code{\link{hopit.control}}.}
}
\description{
Fit the model.
}
\author{
Maciej J. Danko
}
\keyword{internal}
