% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hopit.R
\name{hopit_Latent}
\alias{hopit_Latent}
\title{INTERNAL: Calculate latent variable}
\usage{
hopit_Latent(reg.params, model = NULL)
}
\description{
INTERNAL: Calculate latent variable
}
\author{
Maciej J. Danko
}
\keyword{internal}
