% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hopit.R
\name{hopit_latentrange}
\alias{hopit_latentrange}
\title{INTERNAL: Calculate maximum possible latent range}
\usage{
hopit_latentrange(model, data)
}
\arguments{
\item{model}{a fitted \code{hopit} model.}

\item{data}{a data used to fit the model}
}
\description{
INTERNAL: Calculate maximum possible latent range
}
\keyword{internal}
