% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hopit.R
\name{hopit_negLL}
\alias{hopit_negLL}
\title{INTERNAL: The log likelihood function}
\usage{
hopit_negLL(parameters = model$coef, model, collapse = TRUE,
  use_weights = TRUE, negative = TRUE)
}
\description{
INTERNAL: The log likelihood function
}
\author{
Maciej J. Danko
}
\keyword{internal}
