% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hopit.R
\name{hopit_Threshold}
\alias{hopit_Threshold}
\title{INTERNAL: Calculation of cut-points (threshold)}
\usage{
hopit_Threshold(thresh.lambda, thresh.gamma, model)
}
\description{
INTERNAL: Calculation of cut-points (threshold)
}
\author{
Maciej J. Danko
}
\keyword{internal}
