% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportingstyles.R
\name{latentIndex}
\alias{latentIndex}
\alias{healthIndex}
\title{Calculate latent index}
\usage{
latentIndex(model, decreasing.levels = TRUE, subset = NULL, plotf = FALSE,
  response = c("data", "fitted", "Jurges"), ylab = "Latent index", ...)

healthIndex(model, decreasing.levels = TRUE, subset = NULL, plotf = FALSE,
  response = c("data", "fitted", "Jurges"), ylab = "Latent index", ...)
}
\arguments{
\item{model}{a fitted \code{hopit} model.}

\item{decreasing.levels}{logical indicating if self-reported (e.g. health) classes are ordered in decreasing order.}

\item{subset}{an optional vector specifying a subset of observations.}

\item{plotf}{logical indicating if to plot summary figure.}

\item{response}{X axis plotting option, choose \code{'data'} for raw responses and \code{'fitted'} for model reclassified responses}

\item{ylab}{a label of y axis.}

\item{...}{futher parameters passedto the \code{\link{plot}} function.}
}
\description{
Calcualte latent index from the latent variable. It takes values from 0 to 1, where
zero is prescribed to the worse predicted state (maximal observed value fo the latent variable) and 1 is prescribed
to the best predicted health (minimal observed value fo the latent variable).
}
\author{
Maciej J. Danko
}
