% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportingstyles.R
\name{latentindexStats}
\alias{latentindexStats}
\alias{healthindexStats}
\title{L}
\usage{
latentindexStats(model, formula, data = environment(model$thresh.formula),
  plotf = TRUE, sep = "\\n", sort.flag = FALSE, mar = c(4, 8, 1.5, 0.5),
  oma = c(0, 0, 0, 0), XLab = "Health index",
  healthlevelsorder = "decreasing")

healthindexStats(model, formula, data = environment(model$thresh.formula),
  plotf = TRUE, sep = "\\n", sort.flag = FALSE, mar = c(4, 8, 1.5, 0.5),
  oma = c(0, 0, 0, 0), XLab = "Health index",
  healthlevelsorder = "decreasing")
}
\arguments{
\item{model}{a fitted \code{hopit} model.}

\item{formula}{a formula containing the variables. It is by default set to threshold formula.}

\item{data}{used to fit the model.}

\item{plotf}{logical indicating if to plot the results.}

\item{sep}{separator for levls names.}

\item{mar}{see \code{\link{par}}.}

\item{oma}{see \code{\link{par}}.}

\item{healthlevelsorder}{order of self-reported healh levels. Possible values are \code{'increasing'} and \code{'decreasing'}}
}
\description{
Calculate and plot interquartile ranges, medians and means for latent index for selected grouping variables
}
