% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gotm.r
\name{logLik.gotm}
\alias{logLik.gotm}
\title{Extracts log likelihood of the fitted model}
\usage{
logLik.gotm(object, ...)
}
\arguments{
\item{object}{\code{gotm} object.}

\item{...}{additional objects of the same type.}
}
\description{
Extracts log likelihood of the fitted model
}
\author{
Maciej J. Danko <\email{danko@demogr.mpg.de}> <\email{maciej.danko@gmail.com}>
}
\keyword{internal}

