% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gotm.r
\name{lrt.gotm}
\alias{lrt.gotm}
\title{Likelihood ratio test for a pair of models}
\usage{
lrt.gotm(full, nested)
}
\arguments{
\item{full, nested}{{Models to be compared.}}
}
\description{
Likelihood ratio test for a pair of models
}
\author{
Maciej J. Danko <\email{danko@demogr.mpg.de}> <\email{maciej.danko@gmail.com}>
}
\keyword{internal}

