% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{lrt.hopit}
\alias{lrt.hopit}
\title{Likelihood ratio test for a pair of models}
\usage{
lrt.hopit(full, nested)
}
\arguments{
\item{full, nested}{Models to be compared.}
}
\description{
Likelihood ratio test for a pair of models
}
\author{
Maciej J. Danko
}
\keyword{internal}
