% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{my.grad}
\alias{my.grad}
\title{INTERNAL: numerical gradient}
\usage{
my.grad(fn, par, eps, ...)
}
\arguments{
\item{fn}{function}

\item{par}{parameters}

\item{eps}{epsilon}

\item{...}{other parameters passed to fn}
}
\description{
INTERNAL: numerical gradient
}
\author{
Maciej J. Danko
}
\keyword{internal}
