% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot.R
\name{percentile_CI}
\alias{percentile_CI}
\title{Calculating confidence intervals of bootstrapped function using percentile method}
\usage{
percentile_CI(boot, alpha = 0.05, bounds = c("both", "lo", "up"))
}
\arguments{
\item{boot}{a matrix or list of vectors with bootstrapped elements. If a list then each element of the list is one replication.}

\item{alpha}{significance level.}

\item{bounds}{one of \code{"both"}, \code{"lo"}, \code{"up"}.}
}
\description{
Calculating confidence intervals of bootstrapped function using percentile method
}
\examples{
# see examples in boot_hopit() function.
}
\seealso{
\code{\link{boot_hopit}}, \code{\link{getLevels}}, \code{\link{getCutPoints}}, \code{\link{latentIndex}}, \code{\link{standardiseCoef}}, \code{\link{hopit}}.
}
\author{
Maciej J. Danko
}
