% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{plot.hopitCP}
\alias{plot.hopitCP}
\title{Plotting Cut-Points}
\usage{
\method{plot}{hopitCP}(x, decreasing.levels = x$decreasing.levels,
  mar = c(4, 4, 1, 1), oma = c(0, 0, 0, 0), XLab = "Health index",
  XLab.cex = 1.1, YLab = "Counts", YLab.cex = 1.1, border.col = 2,
  border.lty = 2, border.lwd = 1.5, group.labels.type = c("middle",
  "border", "none"), ...)
}
\arguments{
\item{decreasing.levels}{a logical indicating whether self-reported health classes are ordered in decreasing order.}

\item{mar, oma}{graphic parameters, see \code{\link{par}}.}

\item{XLab, XLab.cex}{a label of the x axis and it's size.}

\item{YLab, YLab.cex}{a label of the y axis and it's size.}

\item{border.lwd, border.lty, border.col}{graphic parameters for vertical lines used to plot cut-points.}

\item{group.labels.type}{a position of the legend. One of \code{middle}, \code{border}, or \code{none}.}

\item{...}{further plotting arguments.}

\item{plotf}{a logical indicating whether to plot the results.}
}
\description{
Plotting Cut-Points
}
\author{
Maciej J. Danko
}
\keyword{internal}
