% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{plot.hopitDW}
\alias{plot.hopitDW}
\title{Plotting standardized coefficients}
\usage{
\method{plot}{hopitDW}(x, ordered = TRUE, show.signif = TRUE,
  mar = c(10, 4, 1.5, 1), oma = c(0, 0, 0, 0),
  ylab = "Disability weight", xlab = "", density = 20, angle = 45,
  col = "orange", las = 3, ...)
}
\arguments{
\item{x}{a object generated by \code{\link{standardizeCoef}} function.}

\item{ordered}{a logical indicating whether to sort the disability weights.}

\item{show.signif}{show significance codes.}

\item{mar, oma}{graphic parameters, see \code{\link{par}}.}

\item{ylab, xlab, density, angle, col, las, ...}{arguments passed to \code{\link{barplot}}.}
}
\description{
Plotting standardized coefficients
}
\author{
Maciej J. Danko
}
\keyword{internal}
