% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{plot.hopitHI}
\alias{plot.hopitHI}
\title{Plotting Latent Index}
\usage{
\method{plot}{hopitHI}(x, response = c("data", "fitted"), xlab = "",
  ylab = "Latent index", ...)
}
\arguments{
\item{x}{a object generated by \code{\link{latentIndex}} function.}

\item{response}{X-axis plotting option; choose \code{'data'} for the raw responses and \code{'fitted'} for the responses reclassified by the model.}

\item{xlab}{a label of the x-axis.}

\item{ylab}{a label of the y-axis.}

\item{...}{further parameters passed to the \code{\link{plot}} function.}
}
\description{
Plotting Latent Index
}
\author{
Maciej J. Danko
}
\keyword{internal}
