% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{plot.hopitLV}
\alias{plot.hopitLV}
\title{Plotting getLevels object}
\usage{
\method{plot}{hopitLV}(x, mar = c(7, 2, 1.5, 0.5), oma = c(0, 3, 0, 0),
  YLab = "Fraction [\%]", YLab.cex = 1.1,
  legbg = grDevices::adjustcolor("white", alpha.f = 0.4), legbty = "o",
  las = 3, ...)
}
\arguments{
\item{x}{aobject generated by \code{\link{getLevels}}.}

\item{mar, oma}{graphic parameters, see \code{\link{par}}.}

\item{YLab, YLab.cex}{a label for the y-axis and it's size.}

\item{legbg}{a legend background color. See \code{bg} parameter in \code{\link{legend}}.}

\item{legbty}{a legend box type. See \code{bty} parameter in \code{\link{legend}}.}

\item{las}{the style of axis labels, see \code{\link{par}}.}

\item{...}{further plotting arguments.}
}
\description{
Plotting getLevels object
}
\author{
Maciej J. Danko
}
\keyword{internal}
