% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{predict.hopit}
\alias{predict.hopit}
\title{Model predictions}
\usage{
\method{predict}{hopit}(object, newdata=NULL,
type = c('link', 'response', 'threshold', 'threshold_link'),
unravelFreq = TRUE, ...)
}
\arguments{
\item{object}{\code{hopit} object.}

\item{type}{the type of prediction required. The default \code{"link"}
is on the scale of linear predictors (latent variable). The alternative \code{"response"}
is on the scale of categorical response variable. The \code{"threshold"}
gives the thresholds for each observation, whereas the \code{"threshold_link"} gives meaningful thresholds
together with latent variable for each observation (a data.frame with fields \code{$left.boundary},
\code{$latent.variable}, and \code{$right.boundary}).}

\item{unravelFreq}{logical indicating if to represent results on individual scale if FWeights were used.}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Model predictions
}
\author{
Maciej J. Danko
}
