% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gotm.r
\name{print.anova.gotm}
\alias{print.anova.gotm}
\title{Print object calcuated by \code{\link{anova.gotm}}}
\usage{
\method{print}{anova.gotm}(x, ...)
}
\arguments{
\item{x}{object generated by \code{\link{anova.gotm}}}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Print object calcuated by \code{\link{anova.gotm}}
}
\author{
Maciej J. Danko <\email{danko@demogr.mpg.de}> <\email{maciej.danko@gmail.com}>
}
\keyword{internal}

