% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{print.hopit}
\alias{print.hopit}
\title{Printing basic information about fitted hopit}
\usage{
print.hopit(x, ...)
}
\arguments{
\item{x}{\code{hopit} object.}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Printing basic information about fitted hopit
}
\author{
Maciej J. Danko
}
\keyword{internal}
