% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{print.lrt.gotm}
\alias{print.lrt.gotm}
\title{Print object calculated by \code{\link{lrt.gotm}}}
\usage{
print.lrt.gotm(x, ...)
}
\arguments{
\item{x}{object obtained from \code{\link{lrt.gotm}}}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Print object calculated by \code{\link{lrt.gotm}}
}
\author{
Maciej J. Danko
}
\keyword{internal}
