% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{print.lrt.hopit}
\alias{print.lrt.hopit}
\title{Print object calculated by \code{\link{lrt.hopit}}}
\usage{
\method{print}{lrt.hopit}(x, ...)
}
\arguments{
\item{x}{object obtained from \code{\link{lrt.hopit}}}

\item{short}{logical, shortened description}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Print object calculated by \code{\link{lrt.hopit}}
}
\author{
Maciej J. Danko
}
\keyword{internal}
