% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{print.summary.gotm}
\alias{print.summary.gotm}
\title{Print object calculated by \code{\link{summary.gotm}}}
\usage{
print.summary.gotm(x, ...)
}
\arguments{
\item{x}{object created with \code{\link{summary.gotm}}}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Print object calculated by \code{\link{summary.gotm}}
}
\author{
Maciej J. Danko
}
\keyword{internal}
