% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gotm.r
\name{print.vcov.gotm}
\alias{print.vcov.gotm}
\title{Print object calculated by \code{\link{vcov.gotm}}}
\usage{
\method{print}{vcov.gotm}(x, digits = 3L, ...)
}
\arguments{
\item{x}{\code{gotm} object}
}
\description{
Print object calculated by \code{\link{vcov.gotm}}
}
\keyword{internal}

