% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{print.vcov.hopit}
\alias{print.vcov.hopit}
\title{Print object calculated by \code{\link{vcov.hopit}}}
\usage{
\method{print}{vcov.hopit}(x, digits = 3L, ...)
}
\arguments{
\item{x}{\code{hopit} object}

\item{digits}{see \code{\link{print.default}}}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Print object calculated by \code{\link{vcov.hopit}}
}
\keyword{internal}
