% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gotm.r
\name{resid.gotm}
\alias{resid.gotm}
\alias{residuals}
\title{Extracting residuals from the fitted gotm}
\usage{
\method{resid}{gotm}(object, type = c("working", "response"), ...)
}
\arguments{
\item{object}{\code{gotm} object.}

\item{type}{type of residuals}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Extracting residuals from the fitted gotm
}
\author{
Maciej J. Danko <\email{danko@demogr.mpg.de}> <\email{maciej.danko@gmail.com}>
}
\keyword{internal}

