% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{sigma.hopit}
\alias{sigma.hopit}
\title{Extract the \code{Sigma} parameter from a \code{hopit} model}
\usage{
\method{sigma}{hopit}(object, ...)
}
\arguments{
\item{model}{a fitted \code{hopit} model.}
}
\description{
Extract the \code{Sigma} parameter from a \code{hopit} model
}
\author{
Maciej J. Danko
}
\keyword{internal}
