% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportingheterogeneity.R
\name{standardizeCoef}
\alias{standardizeCoef}
\alias{standardiseCoef}
\alias{disabilityWeights}
\title{Standardization of coefficients}
\usage{
standardizeCoef(model, ordered = TRUE, plotf = FALSE,
  plotpval = FALSE, mar = c(15, 4, 1, 1), oma = c(0, 0, 0, 0),
  YLab = "Disability weight", YLab.cex = 1.1, namesf = identity, ...)

standardiseCoef(model, ordered = TRUE, plotf = FALSE,
  plotpval = FALSE, mar = c(15, 4, 1, 1), oma = c(0, 0, 0, 0),
  YLab = "Disability weight", YLab.cex = 1.1, namesf = identity, ...)

disabilityWeights(model, ordered = TRUE, plotf = FALSE,
  plotpval = FALSE, mar = c(15, 4, 1, 1), oma = c(0, 0, 0, 0),
  YLab = "Disability weight", YLab.cex = 1.1, namesf = identity, ...)
}
\arguments{
\item{model}{a fitted \code{hopit} model.}

\item{ordered}{logical indicating if to order the disability weights.}

\item{plotf}{logical indicating if to plot results.}

\item{plotpval}{logical indicating if to plot p-values.}

\item{mar, oma}{see \code{\link{par}}.}

\item{YLab, YLab.cex}{label and \code{cex} of y axis.}

\item{namesf}{a vector of names of coefficients or one argument function that modifies names of coefficients.}

\item{...}{arguments passed to \code{\link{boxplot}}.}
}
\value{
a vector with standardized coefficients.
}
\description{
Calculate standardized coefficients (e.g. disability weights for health variables) using
the predicted latent measure obtained from the model.\cr
In the self-rated health example the standardized coefficients are called disability weights \insertCite{Jurges2007;textual}{hopit}
and are calculated for each health variable to provide information about the impact of a specific health measure on the latent index
(see \code{\link{latentIndex}}). The disability weight for a health variable is equal to the ratio of corresponding health coefficient
and the difference between the lowest and highest values of predicted latent health. In other words, disability weight reduces
latent index by some given amount or percentage (i.e.  of every individual is reduced by the same amount if heart attack or other
heart problems are present)\insertCite{Jurges2007;textual}{hopit}.
}
\examples{
# DATA
data(healthsurvey)

# the order of response levels is decreasing (from the best health to the worst health)
levels(healthsurvey$health)

# Example 1 ---------------------

# fitting a model
model1 <- hopit(latent.formula = health ~ hypertension + high_cholesterol +
                heart_attack_or_stroke + poor_mobility + very_poor_grip +
                depression + respiratory_problems +
                IADL_problems + obese + diabetes + other_diseases,
              thresh.formula = ~ sex + ageclass + country,
              decreasing.levels = TRUE,
              control = list(trace = FALSE),
              data = healthsurvey)

# A function that modifies coefficient names.
txtfun <- function(x) gsub('_',' ',substr(x,1,nchar(x)-3))

# Calculate and plot disability weights
sc <- standardizeCoef(model1, plotf = TRUE, namesf = txtfun)
sc
}
\references{
\insertRef{Jurges2007}{hopit}
}
\seealso{
\code{\link{latentIndex}}, \code{\link{getCutPoints}}, \code{\link{getLevels}}, \code{\link{hopit}}.
}
\author{
Maciej J. Danko
}
