% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{start.vglm}
\alias{start.vglm}
\title{INTERNAL: Use vglm to get starting parameters}
\usage{
\method{start}{vglm}(model, data)
}
\arguments{
\item{model}{\code{hopit} object.}

\item{data}{data.frame with data used to fit the model.}
}
\value{
updated model
importFrom VGAM vglm
}
\description{
INTERNAL: Use vglm to get starting parameters
}
\keyword{internal}
