% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gotm.r
\name{summary.gotm}
\alias{summary.gotm}
\title{Calcualte model summary}
\usage{
\method{summary}{gotm}(object, robust.se = FALSE, control = list(), ...)
}
\arguments{
\item{object}{\code{gotm} object.}

\item{robust.se}{logical indicating if to use robust standard errors based on the sandwich estimator.
If survey deign is detected than this option is ignored.}

\item{control}{a list with control parameters. See \code{\link{gotm.control}}.}
}
\description{
Calcualte model summary
}
\author{
Maciej J. Danko <\email{danko@demogr.mpg.de}> <\email{maciej.danko@gmail.com}>
}
\keyword{internal}

