% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{summary.hopit}
\alias{summary.hopit}
\title{Calculate model summary}
\usage{
summary.hopit(object, robust.se = FALSE, ...)
}
\arguments{
\item{object}{\code{hopit} object.}

\item{robust.se}{logical indicating if to use robust standard errors based on the sandwich estimator.
If survey deign is detected than this option is ignored.}

\item{...}{further arguments passed to or from other methods.}

\item{control}{a list with control parameters. See \code{\link{hopit.control}}.}
}
\description{
Calculate model summary
}
\author{
Maciej J. Danko
}
