% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surveyclone.R
\name{svy.varcoef_hopit}
\alias{svy.varcoef_hopit}
\title{Calculation of the variance-covariance matrix for a specified survey design (experimental function)}
\usage{
svy.varcoef_hopit(Ainv, estfun, design)
}
\arguments{
\item{Ainv}{a variance-covariance matrix.}

\item{estfun}{a gradient function of the log-likelihood function.}

\item{design}{a \code{survey.design} object.}
}
\description{
This is a modification of \code{survey:::svy.varcoef}. In the original approach \code{estfun} is calculated from
glm's working residuals:\cr
\code{estfun <- model.matrix(glm.object) * resid(glm.object, "working") * glm.object$weights}\cr
In the hopit package, estfun is directly calculated as a gradient (vector of partial derivatives) of the log likelihood function.
}
\details{
Based on the survey v3.35 package.
}
\seealso{
\code{\link[survey]{svydesign}}
\code{\link{hopit}}
}
\author{
Thomas Lumley, modified by Maciej J. Danko
}
