% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gotm.r
\name{vcov.gotm}
\alias{vcov.gotm}
\title{#' Extracting residuals from the fitted gotm
#'
#' @param object \code{gotm} object.
#' @param type type of residuals
#' @param ...	further arguments passed to or from other methods.
#' @aliases residuals
#' @keywords internal
#' @author Maciej J. Danko <\email{danko@demogr.mpg.de}> <\email{maciej.danko@gmail.com}>
#' @export
resid.gotm<-function(object, type = c("working", "response"), ...){
  warning("resid.gotm() is an experimentaly function, proceed with caution.")
  Y <- unclass(object$y_i)
  Yhat <- unclass(object$Ey_i)
  res <- Y - Yhat
  type <- tolower(type[1])
  if (!(type %in%  c("working", "response"))) stop('Unknown type.')
  if (type == 'working') {
    stop('Don't know how to calculate)
    #cate <- predict(object, type = 'threshold_link')
    Ygrad <- dnorm(object$y_latent_i)
    res <- res / Ygrad
  }
  res
}
Extracting variance-covariance matrix from the fitted gotm}
\usage{
vcov.gotm(object, robust.vcov, control = list(), ...)
}
\arguments{
\item{object}{\code{gotm} object.}

\item{robust.vcov}{logical indicating if to use sandwich estimator to calculate variance-covariance matrix.
If survey deign is detected than this option is ignored.}

\item{control}{a list with control parameters. See \code{\link{gotm.control}}.}

\item{...}{further arguments passed to or from other methods.}
}
\description{
#' Extracting residuals from the fitted gotm
#'
#' @param object \code{gotm} object.
#' @param type type of residuals
#' @param ...	further arguments passed to or from other methods.
#' @aliases residuals
#' @keywords internal
#' @author Maciej J. Danko <\email{danko@demogr.mpg.de}> <\email{maciej.danko@gmail.com}>
#' @export
resid.gotm<-function(object, type = c("working", "response"), ...){
  warning("resid.gotm() is an experimentaly function, proceed with caution.")
  Y <- unclass(object$y_i)
  Yhat <- unclass(object$Ey_i)
  res <- Y - Yhat
  type <- tolower(type[1])
  if (!(type %in%  c("working", "response"))) stop('Unknown type.')
  if (type == 'working') {
    stop('Don't know how to calculate)
    #cate <- predict(object, type = 'threshold_link')
    Ygrad <- dnorm(object$y_latent_i)
    res <- res / Ygrad
  }
  res
}
Extracting variance-covariance matrix from the fitted gotm
}
\author{
Maciej J. Danko <\email{danko@demogr.mpg.de}> <\email{maciej.danko@gmail.com}>
}

