% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gotm_imprv.r
\name{Vector2DummyMat}
\alias{Vector2DummyMat}
\title{INTERNAL: Converts a vector of an categorical variable into a matrix with dummies in columns}
\usage{
Vector2DummyMat(V)
}
\arguments{
\item{V}{a vector of categories.}
}
\description{
INTERNAL: Converts a vector of an categorical variable into a matrix with dummies in columns
}
\author{
Maciej J. Danko  <\email{maciej.danko@gmail.com}>
}
\keyword{internal}
