% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{AIC.hopit}
\alias{AIC.hopit}
\title{Extracting Akaike Information Criterion from the fitted model}
\usage{
\method{AIC}{hopit}(object, ..., k = 2L)
}
\arguments{
\item{object}{an \code{hopit} object.}

\item{...}{additional objects of the class \code{hopit}.}

\item{k}{a penalty per parameter to be used; the default k = 2 is the
classical AIC.}
}
\description{
Extract Akaike Information Criterion (AIC) from a fitted \code{hopit} model.
}
\author{
Maciej J. Danko
}
\keyword{internal}
