% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot.R
\name{boot_hopit}
\alias{boot_hopit}
\title{Bootstraping hopit model}
\usage{
boot_hopit(model, data, func, nboot = 500, unlist = TRUE, ...)
}
\arguments{
\item{model}{a fitted \code{Hopit} model.}

\item{data}{data used to fit the model.}

\item{func}{function to be bootstrapped of the form \code{func(model, data, ...)}.}

\item{nboot}{number of bootstrap replicates.}

\item{unlist}{logical indicting if to unlist boot object.}

\item{...}{other parameters passed to the \code{func}.}
}
\description{
Bootstraping hopit model
}
\author{
Maciej J. Danko
}
