% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot.R
\name{boot_hopit_CI}
\alias{boot_hopit_CI}
\title{Calculating Confidence Intervals using percentile method}
\usage{
boot_hopit_CI(boot, alpha = 0.05, bounds = c("both", "lo", "up"))
}
\arguments{
\item{boot}{boot object calculated by \code{\link{boot_hopit}} .}

\item{alpha}{significance level.}

\item{bounds}{one of \code{"both"}, \code{"lo"}, \code{"up"}.}
}
\description{
Calculating Confidence Intervals using percentile method
}
\author{
Maciej J. Danko
}
