% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{classify.ind}
\alias{classify.ind}
\title{INTERNAL: Clasify individuals according to the reg.params and calculated thresholds}
\usage{
classify.ind(model)
}
\arguments{
\item{model}{\code{hopit} object.}
}
\description{
INTERNAL: Clasify individuals according to the reg.params and calculated thresholds
}
\keyword{internal}
