% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gotm.R
\name{clone.of.htvar.matrix}
\alias{clone.of.htvar.matrix}
\title{survey:::htvar.matrix clone}
\usage{
clone.of.htvar.matrix(xcheck, Dcheck)
}
\description{
survey:::htvar.matrix clone
}
\author{
Thomas Lumley
}
\keyword{internal}
