% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surveyclone.R
\name{clone.of.ygvar.matrix}
\alias{clone.of.ygvar.matrix}
\title{survey:::ygvar.matrix clone}
\usage{
clone.of.ygvar.matrix(xcheck, Dcheck)
}
\description{
survey:::ygvar.matrix clone
}
\author{
Thomas Lumley
}
\keyword{internal}
