% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gotm.r
\name{coef.gotm}
\alias{coef.gotm}
\alias{coefficients.gotm}
\title{Extracting coefficients of fitted \code{gotm} object}
\usage{
coef.gotm(object, standardized = FALSE, aslist = FALSE, ...)
}
\arguments{
\item{object}{\code{gotm} object.}

\item{standardized}{logical indicating if to standardize the coefficients to get disability weights.}

\item{aslist}{logical indicating if model coefficients should be returned as a list of three vectors
related to latent variable, threshold lambdas, and threshold gammas.}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Extracting coefficients of fitted \code{gotm} object
}
\author{
Maciej J. Danko <\email{danko@demogr.mpg.de}> <\email{maciej.danko@gmail.com}>
}
\keyword{internal}

