% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{coef.hopit}
\alias{coef.hopit}
\alias{coefficients.hopit}
\title{Extracting the model coefficients}
\usage{
\method{coef}{hopit}(object, aslist = FALSE, ...)
}
\arguments{
\item{object}{a \code{hopit} object.}

\item{aslist}{a logical indicating whether the model coefficients should be returned
as a list of three vectors
related to latent variables, threshold lambdas, and threshold gammas.}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Extract the model coefficients from the \code{hopit} model.
}
\author{
Maciej J. Danko
}
\keyword{internal}
