% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gotm.r
\name{col_path}
\alias{col_path}
\title{INTERNAL: Column path in a matrix}
\usage{
col_path(mat, y)
}
\arguments{
\item{mat}{a matrix}

\item{y}{a vector with column indices corresponding to each row in the matrix \code{mat}.}
}
\description{
For each row of a matrix \code{mat} extracts a value corresponding to a column stored in vector \code{y}.
}
\author{
Maciej J. Danko <\email{danko@demogr.mpg.de}> <\email{maciej.danko@gmail.com}>
}
\keyword{internal}

