% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cumsum_row}
\alias{cumsum_row}
\title{INTERNAL: Do cumsum() in each row of a matrix}
\usage{
cumsum_row(mat)
}
\arguments{
\item{mat}{a matrix.}
}
\description{
INTERNAL: Do cumsum() in each row of a matrix
}
\keyword{internal}
