% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportingstyles.R
\name{disabilityweights}
\alias{disabilityweights}
\title{Calculate disability weights}
\usage{
disabilityweights(model, latent.method = c("observed", "theoretical"),
  plotpval = FALSE, plotf = TRUE, mar = c(15, 4, 1, 1), oma = c(0, 0, 0,
  0), namesf = identity)
}
\arguments{
\item{model}{a fitted \code{gotm} model.}

\item{plotf}{logical indicating if to plot results.}

\item{mar}{see \code{\link{par}}.}

\item{oma}{see \code{\link{par}}.}
}
\description{
Calculate disability weights
computed as the regression parameters from the generalised ordered probit model divided by the maximum possible range
of its linear prediction. The range is calcualted as difference between maximum and minimum possible value of the latent variable
given estimated parameters.
}
\seealso{
\code{\link{healthindex}}
}
