% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fit.vglm}
\alias{fit.vglm}
\title{INTERNAL: Fit vglm to get parameters of the model}
\usage{
fit.vglm(model, data)
}
\arguments{
\item{model}{\code{gotm} object.}

\item{data}{data.frame with data used to fit the model.}
}
\value{
updated model
importFrom VGAM vglm
}
\description{
INTERNAL: Fit vglm to get parameters of the model
}
\keyword{internal}
