% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get.hopit.start}
\alias{get.hopit.start}
\title{INTERNAL: Translate glm() to hopit() start parameters}
\usage{
get.hopit.start(object, data)
}
\arguments{
\item{object}{\code{hopit} object.}

\item{data}{data.frame with data used to fit the object.}
}
\value{
updated \code{hopit} object.
}
\description{
INTERNAL: Translate glm() to hopit() start parameters
}
\author{
Maciej J. Danko
}
\keyword{internal}
