% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportingstyles.R
\name{getCutPoints}
\alias{getCutPoints}
\title{Calcualte threshold cut-points using Jurges' method}
\usage{
getCutPoints(model, subset = NULL, plotf = TRUE, mar = c(4, 4, 1, 1),
  oma = c(0, 0, 0, 0), XLab = "Health index", XLab.cex = 1.1,
  YLab = "Counts", YLab.cex = 1.1, decreasing.levels = TRUE,
  group.labels.type = c("middle", "border", "none"))
}
\arguments{
\item{model}{a fitted \code{hopit} model.}

\item{subset}{an optional vector specifying a subset of observations.}

\item{plotf}{logical indicating if to plot the results.}

\item{mar, oma}{see \code{\link{par}}.}

\item{XLab, XLab.cex}{label and size of the label for x axis.}

\item{YLab, YLab.cex}{label and size of the label for y axis.}

\item{decreasing.levels}{logical indicating if self-reported health classes are ordered in decreasing order.}

\item{group.labels.type}{position of the legend. One of \code{middel}, \code{border}, or \code{none}.}
}
\value{
a list with following components:
\item{cutpoints}{ cutpoints for adjusted categorical response levels with corresponding percentiles of latent index.}
\item{adjused.levels}{ adjusted categorical response levels for each individual.}
}
\description{
Calcualte threshold cut-points using Jurges' method.
}
\author{
Maciej J. Danko
}
