% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportingstyles.r
\name{getcutpoints}
\alias{getcutpoints}
\title{Calcualte threshold cut-points using Jurges' method}
\usage{
getcutpoints(model, formula = model$thresh.formula,
  data = environment(model$thresh.formula), plotf = TRUE, sep = "_",
  revf = NULL, mar, oma)
}
\arguments{
\item{model}{fitted gotm model.}

\item{formula}{a formula containing the variables. 
It is by default set to threshold formula. 
If set to \code{NULL} then threshold cut-point are calcualted for the whole population.}

\item{data}{data used to fit the model.}

\item{plotf}{logical indicating if to plot the results.}

\item{sep}{separator for levels names.}

\item{revf}{logical indicating if self-reported health classes are ordered in increasing order.}

\item{mar}{see \code{\link(par)}}

\item{oma}{see \code{\link(par)}}
}
\description{
Calcualte threshold cut-points using Jurges' method.
}

