% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportingstyles.r
\name{gethealthindexquantiles}
\alias{gethealthindexquantiles}
\title{Get health index quantiles with respect to specified vaiables}
\usage{
gethealthindexquantiles(model, formula = model$thresh.formula,
  data = environment(model$thresh.formula), plotf = TRUE, sep = "_",
  mar = c(4, 8, 1.5, 0.5), oma = c(0, 0, 0, 0))
}
\arguments{
\item{model}{fitted gotm model.}

\item{formula}{a formula containing the variables. It is by default set to threshold formula.}

\item{data}{used to fit the model.}

\item{plotf}{logical indicating if to plot the results.}

\item{sep}{separator for levls names.}

\item{mar}{see \code{\link{par}}}

\item{oma}{see \code{\link{par}}}
}
\description{
Get health index quantiles with respect to specified vaiables.
}

