% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportingstyles.R
\name{getLevels}
\alias{getLevels}
\title{Summarize adjusted and original response levels.}
\usage{
getLevels(model, formula = model$thresh.formula,
  data = environment(model$thresh.formula), decreasing.levels = TRUE,
  sort.flag = FALSE, plotf = TRUE, sep = "_", mar = c(7, 2, 1.5, 0.5),
  oma = c(0, 3, 0, 0), YLab = "Fraction [\%]", YLab.cex = 1.1,
  legbg = adjustcolor("white", alpha.f = 0.4), legbty = "o")
}
\arguments{
\item{model}{a fitted \code{hopit} model.}

\item{formula}{a formula containing the grouping variables. It is by default set to threshold formula.}

\item{data}{data used to fit the model.}

\item{decreasing.levels}{logical indicating if self-reported health classes are ordered in increasing order.}

\item{sort.flag}{logical indicating if to sort the levels.}

\item{plotf}{logical indicating if to plot the results.}

\item{sep}{separator for levels names.}

\item{mar, oma}{see \code{\link{par}}.}

\item{YLab, YLab.cex}{label and size of the label for y axis.}

\item{legbg}{legend background color. See \code{bg} parameter in \code{\link{legend}}.}

\item{legbty}{legend box type. See \code{bty} parameter in \code{\link{legend}}.}
}
\value{
a list with following components:
\item{original}{ .}
\item{adjused}{ .}
\item{N.original}{ .}
\item{N.adjused}{ .}
\item{I.original}{ .}
\item{I.adjused}{ .}
\item{tab}{ .}
\item{mat}{ .}
}
\description{
Summarize adjusted and original response levels
}
\author{
Maciej J. Danko
}
