% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hopit.R
\name{gettheta}
\alias{gettheta}
\title{Extract Theta parameter from the model}
\usage{
gettheta(model)
}
\arguments{
\item{moodel}{fitted \code{hopit} model.}
}
\description{
Extract Theta parameter from the model
}
\author{
Maciej J. Danko
}
