% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hopit.R
\name{getTheta}
\alias{getTheta}
\title{Extract Theta parameter from the \code{hopit} model}
\usage{
getTheta(model)
}
\arguments{
\item{model}{a fitted \code{hopit} model.}
}
\description{
Extract Theta parameter from the \code{hopit} model
}
\author{
Maciej J. Danko
}
