% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gotm_imprv.r
\name{gotm.control}
\alias{gotm.control}
\title{Auxiliary for controlling \code{gotm} fitting}
\usage{
gotm.control(grad.eps = 3e-05, bgfs.maxit = 10000, cg.maxit = 10000,
  nlm.maxit = 150, bgfs.reltol = 5e-10, cg.reltol = 5e-10,
  nlm.gradtol = 1e-07, nlm.steptol = 1e-07, quick.fit = TRUE,
  trace = TRUE, thresh.start = -Inf, thresh.1.exp = FALSE,
  LL_out_val = -Inf)
}
\arguments{
\item{grad.eps}{epsilon for hessian function.}

\item{quick.fit}{logical, if TRUE extensive optimization methods are ignored and only BFGS and CG methods are run.}

\item{trace}{logical, if to trace model fitting}

\item{thresh.start}{internal parameter (threshold 0), under developement, do not change.}

\item{thresh.1.exp}{internal parameter (logical if to exponentiate threshold function at threshold 1), under developement, do not change.}

\item{LL_out_val}{internal parameter (LL value returned if LL cannot be computed), under developement, do not change.}
}
\description{
Auxiliary function for controlling \code{gotm} fitting. Use this function to set control
parameters of the \code{\link{gotm}} and other related functions.
}
\seealso{
\code{\link{gotm}}
}
\author{
Maciej J. Danko
}
