% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gotm_imprv.r
\name{gotm.design}
\alias{gotm.design}
\title{Auxiliary for setting a simple survey design for \code{gotm}}
\usage{
gotm.design(PWeights = NULL, FWeights = NULL, PSU = NULL,
  CountryID = NULL)
}
\arguments{
\item{PWeights}{Probability weights (the inverse of Probability of an observation being selected into the sample)}

\item{FWeights}{Frequency weights.}

\item{PSU}{Identificator of the PSU unit. Each P- and F- weight should correspond to exactly one PSU.}

\item{CountryID}{PSU are typically prescribed to only one country.}
}
\description{
Auxiliary for setting a simple survey design for \code{gotm}
}
\author{
Maciej J. Danko  <\email{maciej.danko@gmail.com}>
}
