% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gotm.r
\name{gotm}
\alias{gotm}
\title{Fit Generelaized Ordered Choice Threshold Model}
\usage{
gotm(reg.formula, thresh.formula = as.formula("~ 1"), data, survey = list(),
  link = c("probit", "logit"), thresh.method = c("classic", "hopit"),
  start = NULL, doFit = c("full", "vglm", "no"), hessian = TRUE,
  control = list())
}
\arguments{
\item{reg.formula}{formula used to model latent process.}

\item{thresh.formula}{formula used to model threshold variable.
Any dependent variable (left side of "~") will be ignored.}

\item{data}{a data frame including all modeled variables.}

\item{survey}{an optional survey a survey design. Empty list indicates no survey design. See \code{\link{gotm.design}}.}

\item{link}{the link function. The possible values are \code{"probit"} (default) and \code{"logit"}.}

\item{thresh.method}{\itemize{ Define how the threshold is calculated:
\item{\code{"classic"} - Assumes that: \code{alpha(0) = -Inf}, \code{alpha(1)} is a
linear function of threshold covariates + \code{lambda}, \code{alpha(2..J-1) = thresh.fun(lambda, gamma, threshold covariates)},
and \code{alpha(J) = Inf}. See also [1]. Default option.}
\item{\code{"hopit"} - Assumes that: \code{alpha(0) = 0},  \code{alpha(1..J-1) = thresh.fun(lambda, gamma, threshold covarites)},
and \code{alpha(J) = Inf}. See also [2].}
}}

\item{start}{starting values in the form \code{c(latent_parameters, threshold_lambdas, threshold_gammas)}}

\item{doFit}{character string, \code{'full'} perform the full fit, \code{'vglm'} use starting values obteined from vglm,}

\item{control}{a list with control parameters. See \code{\link{gotm.control}}.}

\item{hessain}{logical indicating if to calculate hessian matrix
but will not improve the fit, and \code{'no'} use external starting values, which must be delivered.}
}
\description{
Fit Generelaized Ordered Choice Threshold Model
}
\author{
Maciej J. Danko <\email{danko@demogr.mpg.de}> <\email{maciej.danko@gmail.com}>
}

