% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gotm_imprv.r
\name{gotm}
\alias{gotm}
\title{Fit Generelaized Ordered Choice Threshold Model}
\usage{
gotm(reg.formula, thresh.formula = as.formula("~ 1"), thresh.start = -Inf,
  thresh.1.exp = FALSE, data, survey = list(), link = c("probit",
  "logit"), start = NULL, hessian = TRUE, control = list())
}
\arguments{
\item{reg.formula}{formula used to model latent process.}

\item{thresh.formula}{formula used to model threshold variable.
Any dependent variable (left side of "~") will be ignored.}

\item{data}{a data frame including all modeled variables.}

\item{survey}{an optional survey a survey design. Empty list indicates no survey design. See \code{\link{gotm.design}}.}

\item{link}{the link function. The possible values are \code{"probit"} (default) and \code{"logit"}.}

\item{start}{starting values in the form \code{c(latent_parameters, threshold_lambdas, threshold_gammas)}}

\item{control}{a list with control parameters. See \code{\link{gotm.control}}.}

\item{hessain}{logical indicating if to calculate hessian matrix
but will not improve the fit, and \code{'no'} use external starting values, which must be delivered.}
}
\description{
-- use_weights is unused....
}
\author{
Maciej J. Danko
}
