% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gotm.r
\name{gotm_derivLL}
\alias{gotm_derivLL}
\title{INTERNAL: The gradient of the log likelihood function}
\usage{
gotm_derivLL(parameters, model, collapse = TRUE, negative = FALSE)
}
\description{
INTERNAL: The gradient of the log likelihood function
}
\author{
Maciej J. Danko <\email{danko@demogr.mpg.de}> <\email{maciej.danko@gmail.com}>
}
\keyword{internal}

