% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gotm_imprv.r
\name{gotm_ExtractParameters}
\alias{gotm_ExtractParameters}
\title{INTERNAL: Extract model parameters in a form of list}
\usage{
gotm_ExtractParameters(model, parameters)
}
\arguments{
\item{model}{\code{gotm} object}

\item{parameters}{model parameters (optional). If not delivered then taken from \code{model$coef}}
}
\description{
Extract model parameters in a form of a list
}
\author{
Maciej J. Danko  <\email{maciej.danko@gmail.com}>
}
\keyword{internal}
