% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gotm.r
\name{gotm_fitter}
\alias{gotm_fitter}
\title{INTERNAL: Fit \code{gotm}}
\usage{
gotm_fitter(model, start = model$start, control = list())
}
\arguments{
\item{model}{\code{gotm} object}

\item{start}{starting parameters}

\item{control}{a list with control parameters. See \code{\link{gotm.control}}.}
}
\description{
INTERNAL: Fit \code{gotm}
}
\author{
Maciej J. Danko <\email{danko@demogr.mpg.de}> <\email{maciej.danko@gmail.com}>
}
\keyword{internal}

