% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gotm_imprv.r
\name{gotm_Latent}
\alias{gotm_Latent}
\title{INTERNAL: Calculate latent variable}
\usage{
gotm_Latent(reg.params, model = NULL)
}
\description{
INTERNAL: Calculate latent variable
}
\author{
Maciej J. Danko
}
\keyword{internal}
