% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gotm.r
\name{gotm_Latent}
\alias{gotm_Latent}
\title{INTERNAL: Calculate a latent variable}
\usage{
gotm_Latent(reg.params, model = NULL)
}
\description{
INTERNAL: Calculate a latent variable
}
\author{
Maciej J. Danko <\email{danko@demogr.mpg.de}> <\email{maciej.danko@gmail.com}>
}
\keyword{internal}

