% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gotm.r
\name{gotm_negLL}
\alias{gotm_negLL}
\title{INTERNAL: The log likelihood function}
\usage{
gotm_negLL(parameters, model, collapse = TRUE, include.weights = TRUE)
}
\description{
INTERNAL: The log likelihood function
}
\author{
Maciej J. Danko <\email{danko@demogr.mpg.de}> <\email{maciej.danko@gmail.com}>
}
\keyword{internal}

