% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gotm.R
\name{gotm_Threshold}
\alias{gotm_Threshold}
\title{INTERNAL: Calculation of cut-points (threshold)}
\usage{
gotm_Threshold(thresh.lambda, thresh.gamma, model)
}
\description{
INTERNAL: Calculation of cut-points (threshold)
}
\author{
Maciej J. Danko
}
\keyword{internal}
