% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gotm.r
\name{gotm_Threshold}
\alias{gotm_Threshold}
\title{INTERNAL: Calculation of cut-points (threshold)}
\usage{
gotm_Threshold(thresh.lambda, thresh.gamma, model = NULL,
  extended.output = FALSE)
}
\description{
INTERNAL: Calculation of cut-points (threshold)
}
\author{
Maciej J. Danko <\email{danko@demogr.mpg.de}> <\email{maciej.danko@gmail.com}>
}
\keyword{internal}

