% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{\%c\%}
\alias{\%c\%}
\title{Check whether one set contains all elements of another set}
\usage{
x \%c\% y
}
\arguments{
\item{x, y}{numeric vectors}
}
\description{
Check whether the y set contains all elements of set x
}
\author{
Maciej J. Danko
}
