% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportingstyles.r
\name{healthindex}
\alias{healthindex}
\title{Calculate health index}
\usage{
healthindex(model, subset = NULL, plotf = FALSE)
}
\arguments{
\item{model}{a fitted \code{gotm} model.}

\item{subset}{an optional vector specifying a subset of observations.}

\item{plotf}{logical indicating if to plot summary figure.}
}
\description{
Calcualte halth index from the latent variable. It takes values from 0 to 1, where 
zero is prescribed to the worse possible health (maximum posible value fo the latent variable; 
all conditions/diseases with negative effects on health are present) and 1 is prescribed 
to the best possible health (calculated analogically).
}

