% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportingstyles.r
\name{healthindex}
\alias{healthindex}
\title{Calculate health index}
\usage{
healthindex(model, subset = NULL, plotf = FALSE)
}
\arguments{
\item{model}{a fitted \code{gotm} model.}

\item{subset}{an optional vector specifying a subset of observations.}

\item{plotf}{logical indicating if to plot summary figure.}
}
\description{
Calcualte halth index from the latent variable. It takes values from 0 (the worse possible health) to
1 (the best possible health).
}

