% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hopit.R
\name{hopit.control}
\alias{hopit.control}
\title{Auxiliary for controlling the fitting of \code{hopit} model}
\usage{
hopit.control(grad.eps = 3e-05, bgfs.maxit = 10000, cg.maxit = 10000,
  nlm.maxit = 150, bgfs.reltol = 5e-10, cg.reltol = 5e-10,
  nlm.gradtol = 1e-07, nlm.steptol = 1e-07, fit.methods = c("CG",
  "BFGS"), quick.fit = TRUE, trace = TRUE, thresh.start = -Inf,
  thresh.1.exp = FALSE, LL_out_val = -Inf)
}
\arguments{
\item{grad.eps}{epsilon for numerical Hessian function.}

\item{bgfs.maxit, cg.maxit, nlm.maxit}{the maximum number of iterations.
See \code{\link{optim}} and \code{\link{nlm}} for details.}

\item{bgfs.reltol, cg.reltol}{relative convergence tolerance.
See \code{\link{optim}} for details.}

\item{nlm.gradtol, nlm.steptol}{tolerance at which the scaled gradient is
considered close enough to zero and
minimum allowable relative step length. See \code{\link{nlm}}.}

\item{fit.methods}{either 'CG' or 'BFGS'. See \code{\link{optim}}.}

\item{quick.fit}{logical, if TRUE extensive \code{nlm} optimization method
is ignored and only BFGS and CG methods are run.}

\item{trace}{logical, if to trace model fitting.}

\item{LL_out_val, thresh.1.exp, thresh.start}{internal parameters under
development, do not change.}
}
\description{
Auxiliary function for controlling the fitting of \code{hopit} model.
Use this function to set control
parameters of the \code{\link{hopit}} and other related functions.
}
\seealso{
\code{\link{hopit}}
}
\author{
Maciej J. Danko
}
