% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hopit.R
\name{hopit}
\alias{hopit}
\title{Fit Generelaized Ordered Choice Threshold Model}
\usage{
hopit(reg.formula, thresh.formula = as.formula("~ 1"), data,
  method = c("hopit", "vglm"), overdispersion = FALSE, design = list(),
  weights = NULL, link = c("probit", "logit"), start.method = c("glm",
  "vglm"), start = NULL, control = list())
}
\arguments{
\item{reg.formula}{formula used to model latent process.}

\item{thresh.formula}{formula used to model threshold variable.
Any dependent variable (left side of "~") will be ignored.}

\item{data}{a data frame including all modeled variables.}

\item{design}{an optional survey design. Use \code{\link[survey]{svydesign}} function to specify the design.}

\item{link}{the link function. The possible values are \code{"probit"} (default) and \code{"logit"}.}

\item{start}{starting values in the form \code{c(latent_parameters, threshold_lambdas, threshold_gammas)}}

\item{control}{a list with control parameters. See \code{\link{hopit.control}}.}

\item{weigts}{an optional weights. Use design to construct survey weights.}
}
\description{
-- use_weights is unused....
}
\author{
Maciej J. Danko
}
