% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hopit.R
\name{hopit}
\alias{hopit}
\title{Fit generelaized hierarchical ordered threshold models.}
\usage{
hopit(latent.formula, thresh.formula = ~1, strata.formula = ~1, data,
  decreasing.levels, start = NULL, overdispersion = FALSE,
  design = list(), weights = NULL, link = c("probit", "logit"),
  control = list())
}
\arguments{
\item{latent.formula}{formula used to model latent variable.}

\item{thresh.formula}{formula used to model threshold variable.
Any dependent variable (left side of "~") will be ignored.}

\item{strata.formula}{formula used to model "stratified" latent variables.
Each term in this formula will be interacted with all latent variables.
Each term will be added to \code{threshold.formula}, while interaction will be added to the \code{latent.formula}.
Any dependent variable (left side of "~") will be ignored.}

\item{data}{a data frame including all modeled variables.}

\item{decreasing.levels}{logical indicating if self-reported health classes are ordered in decreasing order.}

\item{start}{a vector with starting values in the form \code{c(latent_parameters, threshold_lambdas, threshold_gammas)}.}

\item{overdispersion}{logical indicting if to fit additionalparameter theta modeling a variance of error term.}

\item{design}{an optional survey design. Use \code{\link[survey]{svydesign}} function to specify the design.}

\item{weights}{an optional weights. Use design to construct survey weights.}

\item{link}{the link function. The possible values are \code{"probit"} (default) and \code{"logit"}.}

\item{control}{a list with control parameters. See \code{\link{hopit.control}}.}
}
\description{
Fit generelaized hierarchical ordered threshold models.
}
\author{
Maciej J. Danko
}
