% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hopit.R
\name{hopit_derivLL}
\alias{hopit_derivLL}
\title{INTERNAL: The gradient of the log likelihood function}
\usage{
hopit_derivLL(parameters = model$coef, model, collapse = TRUE,
  use_weights, negative = FALSE)
}
\arguments{
\item{parameters}{model parameters (optional). If the parameters not delivered, they are taken from the \code{model$coef}.}

\item{model}{a \code{hopit} object.}

\item{collapse}{a logical indicating whether to sum individual LL contributions.}

\item{use_weights}{a logical indicating whether to use model weights.}

\item{negative}{a logical indicating whether the function should return negative LL.}
}
\description{
INTERNAL: The gradient of the log likelihood function
}
\author{
Maciej J. Danko
}
\keyword{internal}
