% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hopit.R
\name{hopit_fitter}
\alias{hopit_fitter}
\title{INTERNAL: Fit a \code{hopit} model given the starting parameters}
\usage{
hopit_fitter(model, start = model$start, use_weights)
}
\arguments{
\item{model}{a \code{hopit} object.}

\item{start}{starting parameters.}

\item{use_weights}{a logical indicating whether to use model weights.}
}
\description{
Fit the model.
}
\author{
Maciej J. Danko
}
\keyword{internal}
