% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hopit.R
\name{hopit_Latent}
\alias{hopit_Latent}
\title{INTERNAL: Calculate the predicted continuous latent measure (h_i).}
\usage{
hopit_Latent(latent.params, model = NULL)
}
\arguments{
\item{latent.params}{vectors with model parameters.}

\item{model}{a \code{hopit} object}
}
\description{
INTERNAL: Calculate the predicted continuous latent measure (h_i).
}
\author{
Maciej J. Danko
}
\keyword{internal}
