% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hopit.R
\name{hopit_negLL}
\alias{hopit_negLL}
\title{INTERNAL: The log likelihood function}
\usage{
hopit_negLL(parameters = model$coef, model, collapse = TRUE,
  use_weights, negative = TRUE)
}
\arguments{
\item{parameters}{model parameters (optional). If the parameters are not delivered, they are taken from \code{model$coef}.}

\item{model}{a \code{hopit} object.}

\item{collapse}{a logical indicating whether to sum the individual LL contributions.}

\item{use_weights}{a logical indicating whether to use model weights.}

\item{negative}{a logical indicating whether the function should return negative.}
}
\description{
INTERNAL: The log likelihood function
}
\author{
Maciej J. Danko
}
\keyword{internal}
