% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hopit.R
\name{hopit_Threshold}
\alias{hopit_Threshold}
\title{INTERNAL: Calculate the model cut-points (alpha)}
\usage{
hopit_Threshold(thresh.lambda, thresh.gamma, model)
}
\arguments{
\item{thresh.lambda, thresh.gamma}{vectors with model parameters.}

\item{model}{a \code{hopit} object.}
}
\description{
INTERNAL: Calculate the model cut-points (alpha)
}
\author{
Maciej J. Danko
}
\keyword{internal}
