% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportingstyles.R
\name{latentIndex}
\alias{latentIndex}
\alias{healthIndex}
\title{Calculate latent index}
\usage{
latentIndex(model, subset = NULL, plotf = FALSE, crude = FALSE,
  healthlevelsorder = c("decreasing", "increasing"), response = c("data",
  "fitted"), YLab = "Latent index")

healthIndex(model, subset = NULL, plotf = FALSE, crude = FALSE,
  healthlevelsorder = c("decreasing", "increasing"), response = c("data",
  "fitted"), YLab = "Latent index")
}
\arguments{
\item{model}{a fitted \code{hopit} model.}

\item{subset}{an optional vector specifying a subset of observations.}

\item{plotf}{logical indicating if to plot summary figure.}

\item{crude}{logical indicating if to calculate crude health measure based directly on self reported health levels.}

\item{healthlevelsorder}{order of self-reported healh levels. Possible values are \code{'increasing'} and \code{'decreasing'}}

\item{response}{X axis plotting option, choose \code{'data'} for raw responses and \code{'fitted'} for model reclassified responses}

\item{YLab}{a label of y axis.}
}
\description{
Calcualte latent index from the latent variable. It takes values from 0 to 1, where
zero is prescribed to the worse possible state (maximum posible value fo the latent variable;
all conditions/diseases with negative effects on health are present) and 1 is prescribed
to the best possible health (calculated analogically).
}
