% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportingstyles.R
\name{latentIndex}
\alias{latentIndex}
\alias{healthIndex}
\title{Calculate latent index}
\usage{
latentIndex(model, decreasing.levels = TRUE, subset = NULL,
  plotf = FALSE, response = c("data", "fitted", "Jurges"),
  ylab = "Latent index", ...)

healthIndex(model, decreasing.levels = TRUE, subset = NULL,
  plotf = FALSE, response = c("data", "fitted", "Jurges"),
  ylab = "Latent index", ...)
}
\arguments{
\item{model}{a fitted \code{hopit} model.}

\item{decreasing.levels}{logical indicating if self-reported (e.g. health) classes are ordered in decreasing order.}

\item{subset}{an optional vector specifying a subset of observations.}

\item{plotf}{logical indicating if to plot summary figure.}

\item{response}{X axis plotting option, choose \code{'data'} for raw responses and \code{'fitted'} for model reclassified responses}

\item{ylab}{a label of y axis.}

\item{...}{further parameters passed to the \code{\link{plot}} function.}
}
\value{
a vector with latent index for each individual.
}
\description{
Calculate latent index from the latent variable. It takes values from 0 to 1, where
zero is prescribed to the worse predicted state (maximal observed value for the latent variable) and 1 is prescribed
to the best predicted health (minimal observed value fo the latent variable).
}
\examples{
# DATA
data(healthsurvey)

# the order of response levels is decreasing (from the best health to the worst health)
levels(healthsurvey$health)

# Example 1 ---------------------

# fitting a model
model1 <- hopit(latent.formula = health ~ hypertension + high_cholesterol +
                heart_attack_or_stroke + poor_mobility + very_poor_grip +
                depression + respiratory_problems +
                IADL_problems + obese + diabetes + other_diseases,
              thresh.formula = ~ sex + ageclass + country,
              decreasing.levels = TRUE,
              control = list(trace = FALSE),
              data = healthsurvey)

# calculate health index and plotting reported health status
# vs. health index.
hi <- latentIndex(model1, plotf = TRUE, response = "data",
                  ylab = 'Health index', col='deepskyblue3')

# a simple histogram of the function output
hist(hi)

# calculate health index and plotting adjusted health status (Jurges 2007)
# vs. health index.
latentIndex(model1, plotf = TRUE, response = "Jurges",
                 ylab = 'Health index', col='deepskyblue3')

# calculate health index and plotting predicted health status
# vs. health index.
latentIndex(model1, plotf = TRUE, response = "fitted",
                 ylab = 'Health index', col='deepskyblue3')
}
\references{
\insertRef{Jurges2007}{hopit}
}
\seealso{
\code{\link{standardizeCoef}}, \code{\link{getCutPoints}}, \code{\link{getLevels}}, \code{\link{hopit}}.
}
\author{
Maciej J. Danko
}
