% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{logLik.gotm}
\alias{logLik.gotm}
\title{Extracts log likelihood of the fitted model}
\usage{
logLik.gotm(object, ...)
}
\arguments{
\item{object}{\code{gotm} object.}

\item{...}{additional objects of the same type.}
}
\description{
Extracts log likelihood of the fitted model
}
\author{
Maciej J. Danko
}
\keyword{internal}
