% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{logLik.hopit}
\alias{logLik.hopit}
\title{Extracting a log likelihood of the fitted model}
\usage{
\method{logLik}{hopit}(object, ...)
}
\arguments{
\item{object}{a \code{hopit} object.}

\item{...}{additional objects of the class \code{hopit}.}
}
\description{
Extract the log likelihood of a \code{hopit} model.
}
\author{
Maciej J. Danko
}
\keyword{internal}
