% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot.R
\name{percentile_CI}
\alias{percentile_CI}
\title{Calculating the confidence intervals of the bootstrapped function using the percentile method}
\usage{
percentile_CI(boot, alpha = 0.05, bounds = c("both", "lo", "up"))
}
\arguments{
\item{boot}{a matrix or a list of vectors with bootstrapped elements. If it is list, then each element of the list is one replication.}

\item{alpha}{a significance level.}

\item{bounds}{which bounds to return; one of \code{"both"}, \code{"lo"}, \code{"up"}.}
}
\description{
Calculate the confidence intervals of the bootstrapped function using the percentile method.
}
\examples{
# see examples in boot_hopit() function.
}
\seealso{
\code{\link{boot_hopit}}, \code{\link{getLevels}}, \code{\link{getCutPoints}}, \code{\link{latentIndex}}, \code{\link{standardiseCoef}}, \code{\link{hopit}}.
}
\author{
Maciej J. Danko
}
