% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{plot.profile.hopit}
\alias{plot.profile.hopit}
\title{Plot the log likelihood profile for a profile.hopit object}
\usage{
\method{plot}{profile.hopit}(x, ..., ylim = NULL, relative = FALSE,
leg.cex = 0.85, leg.col = 'blue4')
}
\arguments{
\item{x}{a \code{profile.hopit} object.}

\item{...}{arguments to be passed to the \code{\link{plot}}() function (see
\code{\link{par}}).}

\item{ylim}{see \code{\link{plot}}.}

\item{relative}{a logical indicating whether \code{ylim} on each panel should be the
same (\code{TRUE}) or not (\code{FALSE}).}

\item{leg.cex}{a character expansion factor relative to the current
\code{par("cex")} (see \code{\link{legend}}).}

\item{leg.col}{a color used for the legend text.}
}
\description{
Plot the method for a profile.hopit object.
}
\seealso{
\code{\link{profile.hopit}}, \code{\link{print.profile.hopit}},
\code{\link{hopit}}.
}
\author{
Maciej J. Danko
}
\keyword{internal}
