% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{plot.profile.hopit}
\alias{plot.profile.hopit}
\title{Plot log likelihood profile for profile.hopit object}
\usage{
\method{plot}{profile.hopit}(x, ..., leg.cex = 0.85, leg.col = 'blue4')
}
\arguments{
\item{x}{\code{profile.hopit} object.}

\item{...}{arguments to be passed to \code{\link{plot}}() function (see \code{\link{par}}).}

\item{ylim}{see \code{\link{plot}}}

\item{relative}{logical indicating if \code{ylim} on each panel should be the same (\code{TRUE}) or not (\code{FALSE})}

\item{leg.cex}{character expansion factor relative to current \code{par("cex")} (see \code{\link{legend}}).}

\item{leg.col}{the color used for the legend text.}
}
\description{
Plot method for profile.hopit object.
}
\seealso{
\code{\link{profile.hopit}}, \code{\link{print.profile.hopit}}, \code{\link{hopit}}
}
\author{
Maciej J. Danko
}
\keyword{internal}
