% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{plot.profile.hopit}
\alias{plot.profile.hopit}
\title{Plot log likelihood profile for profile.hopit object}
\usage{
\method{plot}{profile.hopit}(object, leg.cex = 0.9, leg.col = "blue")
}
\arguments{
\item{x}{\code{profile.hopit} object.}
}
\description{
Plot log likelihood profile for profile.hopit object
}
\author{
Maciej J. Danko
}
\keyword{internal}
