% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{print.anova.hopit}
\alias{print.anova.hopit}
\title{Print object calcuated by \code{\link{anova.hopit}}}
\usage{
\method{print}{anova.hopit}(x, ...)
}
\arguments{
\item{x}{object generated by \code{\link{anova.hopit}}}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Print object calcuated by \code{\link{anova.hopit}}
}
\seealso{
\code{\link{anova.hopit}}, \code{\link{hopit}}.
}
\author{
Maciej J. Danko
}
\keyword{internal}
