% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gotm_imprv.r
\name{print.gotm}
\alias{print.gotm}
\title{Printing basic information about fitted gotm}
\usage{
print.gotm(x, ...)
}
\arguments{
\item{x}{\code{gotm} object.}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Printing basic information about fitted gotm
}
\author{
Maciej J. Danko  <\email{maciej.danko@gmail.com}>
}
\keyword{internal}
