% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gotm.r
\name{print.gotm}
\alias{print.gotm}
\title{Printing basic information about fitted gotm}
\usage{
\method{print}{gotm}(x, ...)
}
\arguments{
\item{x}{\code{gotm} object.}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Printing basic information about fitted gotm
}
\author{
Maciej J. Danko <\email{danko@demogr.mpg.de}> <\email{maciej.danko@gmail.com}>
}
\keyword{internal}

