% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{print.hopit}
\alias{print.hopit}
\title{Printing basic information about fitted \code{hopit} model}
\usage{
\method{print}{hopit}(x, ...)
}
\arguments{
\item{x}{a \code{hopit} object.}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Print a \code{hopit} model.
}
\author{
Maciej J. Danko
}
\keyword{internal}
