% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gotm.r
\name{print.lrt.gotm}
\alias{print.lrt.gotm}
\title{Print object calculated by \code{\link{lrt.gotm}}}
\usage{
\method{print}{lrt.gotm}(x, ...)
}
\arguments{
\item{x}{object obtained from \code{\link{lrt.gotm}}}
}
\description{
Print object calculated by \code{\link{lrt.gotm}}
}
\author{
Maciej J. Danko <\email{danko@demogr.mpg.de}> <\email{maciej.danko@gmail.com}>
}
\keyword{internal}

