% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{print.lrt.hopit}
\alias{print.lrt.hopit}
\title{Printing an object calculated by \code{\link{lrt.hopit}}}
\usage{
\method{print}{lrt.hopit}(x, short = FALSE, ...)
}
\arguments{
\item{x}{an object obtained from \code{\link{lrt.hopit}}.}

\item{short}{a logical indicating whether to show a shortened description.}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Printing an object calculated by \code{\link{lrt.hopit}}
}
\seealso{
\code{\link{lrt.hopit}}, \code{\link{anova.hopit}},
\code{\link{hopit}}.
}
\author{
Maciej J. Danko
}
\keyword{internal}
