% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{print.profile.hopit}
\alias{print.profile.hopit}
\title{Print method for a profile.hopit object}
\usage{
\method{print}{profile.hopit}(x, ..., plotf = TRUE)
}
\arguments{
\item{x}{a \code{profile.hopit} object.}

\item{...}{arguments to be passed to the \code{plot}() function (see
\code{\link{plot.profile.hopit}}).}

\item{plotf}{a logical indicating whether to plot the profile.}
}
\description{
Print method for a profile.hopit object
}
\seealso{
\code{\link{profile.hopit}}, \code{\link{plot.profile.hopit}},
\code{\link{hopit}}
}
\author{
Maciej J. Danko
}
\keyword{internal}
