% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{print.profile.hopit}
\alias{print.profile.hopit}
\title{Print method for profile.hopit object}
\usage{
\method{print}{profile.hopit}(object, plot = TRUE)
}
\arguments{
\item{x}{\code{hopit} object.}
}
\description{
Print method for profile.hopit object
}
\author{
Maciej J. Danko
}
\keyword{internal}
