% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{print.profile.hopit}
\alias{print.profile.hopit}
\title{Print method for profile.hopit object}
\usage{
\method{print}{profile.hopit}(x, ..., plotf = TRUE)
}
\arguments{
\item{x}{\code{profile.hopit} object.}

\item{...}{arguments to be passed to \code{plot}() function (see \code{\link{plot.profile.hopit}}).}

\item{plotf}{\code{hopit} object.}
}
\description{
Print method for profile.hopit object
}
\seealso{
\code{\link{profile.hopit}}, \code{\link{plot.profile.hopit}}, \code{\link{hopit}}
}
\author{
Maciej J. Danko
}
\keyword{internal}
