% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gotm.r
\name{print.summary.gotm}
\alias{print.summary.gotm}
\title{Print object calculated by \code{\link{summary.gotm}}}
\usage{
\method{print}{summary.gotm}(x, ...)
}
\arguments{
\item{x}{object created with \code{\link{summary.gotm}}}
}
\description{
Print object calculated by \code{\link{summary.gotm}}
}
\author{
Maciej J. Danko <\email{danko@demogr.mpg.de}> <\email{maciej.danko@gmail.com}>
}
\keyword{internal}

