% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{print.summary.hopit}
\alias{print.summary.hopit}
\title{Print an object calculated by \code{\link{summary.hopit}}}
\usage{
\method{print}{summary.hopit}(x, ...)
}
\arguments{
\item{x}{an object created with \code{\link{summary.hopit}}.}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Print an object calculated by \code{\link{summary.hopit}}
}
\author{
Maciej J. Danko
}
\keyword{internal}
