% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gotm_imprv.r
\name{print.vcov.gotm}
\alias{print.vcov.gotm}
\title{Print object calculated by \code{\link{vcov.gotm}}}
\usage{
print.vcov.gotm(x, digits = 3L, ...)
}
\arguments{
\item{x}{\code{gotm} object}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Print object calculated by \code{\link{vcov.gotm}}
}
\keyword{internal}
