% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{profile.hopit}
\alias{profile.hopit}
\title{Calculate log likelihood profile for fitted hopit model}
\usage{
\method{profile}{hopit}(model, scope = 0.15, steps = 101)
}
\arguments{
\item{x}{\code{hopit} object.}
}
\description{
Calculate log likelihood profile for fitted hopit model
}
\author{
Maciej J. Danko
}
\keyword{internal}
