% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{profile.hopit}
\alias{profile.hopit}
\title{Calculate log likelihood profile for fitted hopit model}
\usage{
\method{profile}{hopit}(fitted, ..., scope = 0.15, steps = 101)
}
\arguments{
\item{fitted}{\code{hopit} object. Fitted model.}

\item{...}{unused now.}

\item{scope}{value (fraction) defining plotting range for a coeficient. The range is \code{c(coef \* (1-scope), coef \* (1+scope))}.}

\item{steps}{at how many equaly spaced points calcualte the log likelihood function for each coeficient.}
}
\description{
Calculate log likelihood profile for fitted hopit model
}
\seealso{
\code{\link{plot.profile.hopit}}, \code{\link{print.profile.hopit}}, \code{\link{hopit}}
}
\author{
Maciej J. Danko
}
\keyword{internal}
