% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportingstyles.R
\name{standardizeCoef}
\alias{standardizeCoef}
\alias{standardiseCoef}
\alias{disabilityWeights}
\title{Standardization of coefficients}
\usage{
standardizeCoef(model, ordered = TRUE, plotf = FALSE, plotpval = FALSE,
  mar = c(15, 4, 1, 1), oma = c(0, 0, 0, 0), YLab = "Disability weight",
  YLab.cex = 1.1, namesf = identity, ...)

standardiseCoef(model, ordered = TRUE, plotf = FALSE, plotpval = FALSE,
  mar = c(15, 4, 1, 1), oma = c(0, 0, 0, 0), YLab = "Disability weight",
  YLab.cex = 1.1, namesf = identity, ...)

disabilityWeights(model, ordered = TRUE, plotf = FALSE, plotpval = FALSE,
  mar = c(15, 4, 1, 1), oma = c(0, 0, 0, 0), YLab = "Disability weight",
  YLab.cex = 1.1, namesf = identity, ...)
}
\arguments{
\item{model}{a fitted \code{hopit} model.}

\item{ordered}{logical indicating if to order the disability weights.}

\item{plotf}{logical indicating if to plot results.}

\item{plotpval}{logical indicating if to plot p-values.}

\item{mar, oma}{see \code{\link{par}}.}

\item{YLab, YLab.cex}{label and size of the label for y axis.}

\item{namesf}{a vector of names of coefficients or one argument function that modifies names of coefficients.}

\item{...}{arguments passed to \code{\link{boxplot}}.}
}
\value{
a vector with standardized coefficients.
}
\description{
Calculate standardized coefficients - disability weights
computed as the latent coefficients from the generalised ordered probit model divided by the maximum possible range
of its linear prediction. The range is calculated as difference between maximum and minimum possible value of the latent variable
given estimated parameters.
}
\examples{
# DATA
data(healthsurvey)

# the order of response levels is decreasing (from the best health to the worst health)
levels(healthsurvey$health)

# Example 1 ---------------------

# fitting a model
model1 <- hopit(latent.formula = health ~ hypertension + high_cholesterol +
                heart_attack_or_stroke + poor_mobility + very_poor_grip +
                depression + respiratory_problems +
                IADL_problems + obese + diabetes + other_diseases,
              thresh.formula = ~ sex + ageclass + country,
              decreasing.levels = TRUE,
              control = list(trace = FALSE),
              data = healthsurvey)

# A function that modifies coefficient names.
txtfun <- function(x) gsub('_',' ',substr(x,1,nchar(x)-3))

# Calculate and plot disability weights
sc <- standardizeCoef(model1, plotf = TRUE, namesf = txtfun)
sc
}
\references{
\insertRef{Jurges2007}{hopit}
}
\seealso{
\code{\link{latentIndex}}, \code{\link{getCutPoints}}, \code{\link{getLevels}}, \code{\link{hopit}}.
}
\author{
Maciej J. Danko
}
