% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportingstyles.R
\name{standardizeCoef}
\alias{standardizeCoef}
\alias{standardiseCoef}
\alias{disabilityWeights}
\title{Standardization of coefficients}
\usage{
standardizeCoef(model, ordered = TRUE, plotf = FALSE, plotpval = FALSE,
  mar = c(15, 4, 1, 1), oma = c(0, 0, 0, 0), YLab = "Disability weight",
  YLab.cex = 1.1, namesf = identity, ...)

standardiseCoef(model, ordered = TRUE, plotf = FALSE, plotpval = FALSE,
  mar = c(15, 4, 1, 1), oma = c(0, 0, 0, 0), YLab = "Disability weight",
  YLab.cex = 1.1, namesf = identity, ...)

disabilityWeights(model, ordered = TRUE, plotf = FALSE, plotpval = FALSE,
  mar = c(15, 4, 1, 1), oma = c(0, 0, 0, 0), YLab = "Disability weight",
  YLab.cex = 1.1, namesf = identity, ...)
}
\arguments{
\item{model}{a fitted \code{hopit} model.}

\item{ordered}{logical indicating if to order the disability weights.}

\item{plotf}{logical indicating if to plot results.}

\item{plotpval}{logical indicating if to plot p-values.}

\item{mar, oma}{see \code{\link{par}}.}

\item{YLab, YLab.cex}{label and size of the label for y axis.}

\item{namesf}{vectorof names of coeficients or one argument function that modifies names of coeficients.}

\item{...}{arguments passed to \code{\link{boxplot}}.}
}
\description{
Calculate standardized coeficients - disability weights
computed as the regression parameters from the generalised ordered probit model divided by the maximum possible range
of its linear prediction. The range is calcualted as difference between maximum and minimum possible value of the latent variable
given estimated parameters.
}
\seealso{
\code{\link{healthIndex}}
}
\author{
Maciej J. Danko
}
