% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportingstyles.R
\name{standardizeCoef}
\alias{standardizeCoef}
\alias{standardiseCoef}
\alias{disabilityweights}
\title{Standardization of coefficients}
\usage{
standardizeCoef(model, latent.method = c("observed", "theoretical"),
  plotpval = FALSE, plotf = TRUE, mar = c(15, 4, 1, 1), oma = c(0, 0, 0,
  0), YLab = "Disability weight", namesf = identity, ...)

standardiseCoef(model, latent.method = c("observed", "theoretical"),
  plotpval = FALSE, plotf = TRUE, mar = c(15, 4, 1, 1), oma = c(0, 0, 0,
  0), YLab = "Disability weight", namesf = identity, ...)

disabilityweights(model, latent.method = c("observed", "theoretical"),
  plotpval = FALSE, plotf = TRUE, mar = c(15, 4, 1, 1), oma = c(0, 0, 0,
  0), YLab = "Disability weight", namesf = identity, ...)
}
\arguments{
\item{model}{a fitted \code{hopit} model.}

\item{plotpval}{logical indicating if to plot p-values.}

\item{plotf}{logical indicating if to plot results.}

\item{mar}{see \code{\link{par}}.}

\item{oma}{see \code{\link{par}}.}

\item{namesf}{one argument function that modifies names of coeficients.}

\item{...}{arguments passed to \code{\link{boxplot}}.}
}
\description{
Calculate standardized coeficients - disability weights
computed as the regression parameters from the generalised ordered probit model divided by the maximum possible range
of its linear prediction. The range is calcualted as difference between maximum and minimum possible value of the latent variable
given estimated parameters.
}
\seealso{
\code{\link{healthindex}}
}
