% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{start.glm}
\alias{start.glm}
\title{INTERNAL: Use glm() to get starting parameters}
\usage{
\method{start}{glm}(model, data)
}
\arguments{
\item{model}{\code{hopit} object.}

\item{data}{data.frame with data used to fit the model.}
}
\value{
updated \code{hopit} model.
}
\description{
INTERNAL: Use glm() to get starting parameters
}
\author{
Maciej J. Danko
}
\keyword{internal}
