% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{summary.gotm}
\alias{summary.gotm}
\title{Calculate model summary}
\usage{
\method{summary}{gotm}(object, robust.se = FALSE, ...)
}
\arguments{
\item{object}{\code{gotm} object.}

\item{robust.se}{logical indicating if to use robust standard errors based on the sandwich estimator.
If survey deign is detected than this option is ignored.}

\item{...}{further arguments passed to or from other methods.}

\item{control}{a list with control parameters. See \code{\link{gotm.control}}.}
}
\description{
Calculate model summary
}
\author{
Maciej J. Danko
}
