% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{summary.hopit}
\alias{summary.hopit}
\title{Calculate the model summary}
\usage{
\method{summary}{hopit}(object, robust.se, ...)
}
\arguments{
\item{object}{a \code{hopit} object.}

\item{robust.se}{a logical indicating whether to use robust standard errors based
on the sandwich estimator.
If a survey design is detected, then this option is ignored.}

\item{...}{further arguments passed to or from other methods.}

\item{control}{a list with control parameters.
See \code{\link{hopit.control}}.}
}
\description{
Summarize a \code{hopit}  model.
}
\author{
Maciej J. Danko
}
\keyword{internal}
