% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surveyclone.R
\name{svy.varcoef.hopit}
\alias{svy.varcoef.hopit}
\title{Calculation of variance-covariance matrix for specified survey design}
\usage{
svy.varcoef.hopit(Ainv, estfun, design)
}
\arguments{
\item{Ainv}{a variance-covariance matrix.}

\item{estfun}{LL gradient function.}

\item{design}{a \code{survey.design} object.}
}
\description{
Calculation of variance-covariance matrix for specified survey design
}
\author{
Thomas Lumley, modified by Maciej J. Danko
}
\keyword{internal}
