% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hopit.R
\name{svy.varcoef.hopit}
\alias{svy.varcoef.hopit}
\title{Calculation of variance-covariance matrix for specified survey design}
\usage{
svy.varcoef.hopit(Ainv, estfun, design)
}
\arguments{
\item{design}{a survey.design object}

\item{object}{a hopit object}
}
\description{
Calculation of variance-covariance matrix for specified survey design
}
\author{
Thomas Lumley, modified by Maciej J. Dańko
}
\keyword{internal}
