% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot.R
\name{update.latent}
\alias{update.latent}
\title{INTERNAL : updating model uses new set of latent variable coefficients}
\usage{
update.latent(model, newregcoef, data, hessian = FALSE)
}
\arguments{
\item{model}{a fitted \code{hopit} model.}

\item{newregcoef}{a new set of latent variable coefficients. Vector of length of first element of model$parcount.}

\item{data}{used to fit original model.}

\item{hessian}{logical indicating if to calculate log likelihood gradient (estfun), Hessian and variance-covariance matrix.}
}
\description{
INTERNAL : updating model uses new set of latent variable coefficients
}
\author{
Maciej J. Danko
}
\keyword{internal}
