% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot.R
\name{update.latent}
\alias{update.latent}
\title{INTERNAL : updating model uses new set of latent variable coeficients}
\usage{
update.latent(model, newregcoef, data, hessian = FALSE)
}
\arguments{
\item{model}{a fitted \code{hopit} model.}

\item{newregcoef}{a new set of latent variable coeficients. Vector of length model$parcount\link{1}.}

\item{data}{used to fit original model.}

\item{hessian}{logical indicating if to calculate estfun, hessian and var-cov matrix.}
}
\description{
INTERNAL : updating model uses new set of latent variable coeficients
}
\keyword{internal}
