% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportingstyles.r
\name{update.latent}
\alias{update.latent}
\title{INTERNAL: Update model according to new parameters}
\usage{
update.latent(model, newregcoef, data, update.hessian = FALSE)
}
\arguments{
\item{model}{a fitted \code{gotm} model.}

\item{newregcoef}{new coeficients for health variable deffined in model$reg.formula}

\item{data}{data used to fit the model}

\item{crude}{logical indicating if to re-calculate hessian, variance-covariance matrix, and estfun.}
}
\description{
INTERNAL: Update model according to new parameters
}
\keyword{internal}
