% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gotm.r
\name{vcov.gotm}
\alias{vcov.gotm}
\title{Extracting variance-covariance matrix from the fitted gotm}
\usage{
\method{vcov}{gotm}(object, robust.vcov, control = list(), ...)
}
\arguments{
\item{object}{\code{gotm} object.}

\item{robust.vcov}{logical indicating if to use sandwich estimator to calculate variance-covariance matrix.
If survey deign is detected than this option is ignored.}

\item{control}{a list with control parameters. See \code{\link{gotm.control}}.}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Extracting variance-covariance matrix from the fitted gotm
}
\author{
Maciej J. Danko <\email{danko@demogr.mpg.de}> <\email{maciej.danko@gmail.com}>
}

