% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gotm.r
\name{vcov.gotm}
\alias{vcov.gotm}
\title{Extracting variance-covariance matrix from the fitted gotm}
\usage{
\method{vcov}{gotm}(object, robust.vcov, control = list(),
  robust.method = c("grad", "working"), ...)
}
\arguments{
\item{object}{\code{gotm} object.}

\item{robust.vcov}{logical indicating if to use sandwich estimator to calculate variance-covariance matrix.
If survey deign is detected than this option is ignored.}

\item{control}{a list with control parameters. See \code{\link{gotm.control}}.}

\item{robust.method}{method of calculation of log-likelihood gradient. 
Set \code{"grad"} (default) for numerical gradient or \code{"working"} 
for the method based on working residuals. 
The latter one is an experimental method so warning will apear.}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Extracting variance-covariance matrix from the fitted gotm
}
\author{
Maciej J. Danko <\email{danko@demogr.mpg.de}> <\email{maciej.danko@gmail.com}>
}

