% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{vcov.gotm}
\alias{vcov.gotm}
\title{Extracting variance-covariance matrix from the fitted gotm}
\usage{
vcov.gotm(object, robust.vcov, ...)
}
\arguments{
\item{object}{\code{gotm} object.}

\item{robust.vcov}{logical indicating if to use sandwich estimator to calculate variance-covariance matrix.
If survey deign is detected than this option is ignored.}

\item{...}{further arguments passed to or from other methods.}

\item{control}{a list with control parameters. See \code{\link{gotm.control}}.}
}
\description{
Extracting variance-covariance matrix from the fitted gotm
}
\author{
Maciej J. Danko
}
