% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{vcov.hopit}
\alias{vcov.hopit}
\title{Extracting variance-covariance matrix from the fitted hopit}
\usage{
\method{vcov}{hopit}(object, robust.vcov, ...)
}
\arguments{
\item{object}{\code{hopit} object.}

\item{robust.vcov}{logical indicating if to use sandwich estimator to calculate variance-covariance matrix.
If survey deign is detected than this option is ignored.}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Extracting variance-covariance matrix from the fitted hopit
}
\author{
Maciej J. Danko
}
\keyword{internal}
